@@ -1,22 +1,23 @@
 name: zefyr
 description: Clean, minimalistic and collaboration-ready rich text editor for Flutter.
-version: 0.3.0
+version: 0.8.0
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 homepage: https://github.com/memspace/zefyr
 
 environment:
-  sdk: '>=2.0.0-dev.58.0 <3.0.0'
+  sdk: '>=2.1.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   collection: ^1.14.6
-  url_launcher: ^4.0.0
-  image_picker: ^0.4.5
+  url_launcher: ^5.0.0
   quill_delta: ^1.0.0-dev.1.0
-  notus: ^0.1.0
+  notus: ^0.1.4
   meta: ^1.1.0
+  quiver_hashcode: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  pedantic: ^1.0.0
