@@ -3,11 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:notus/notus.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import 'editor.dart';
+import 'scope.dart';
 import 'theme.dart';
 import 'toolbar.dart';
 
@@ -83,18 +82,18 @@ class ZefyrButton extends StatelessWidget {
       );
     } else {
       assert(_text != null);
-      var style = _textStyle ?? new TextStyle();
+      var style = _textStyle ?? TextStyle();
       style = style.copyWith(color: iconColor);
       return RawZefyrButton(
         action: action,
-        child: new Text(_text, style: style),
+        child: Text(_text, style: style),
         color: _getColor(editor, toolbarTheme),
         onPressed: _getPressedHandler(editor, toolbar),
       );
     }
   }
 
-  Color _getColor(ZefyrEditorScope editor, ZefyrToolbarTheme theme) {
+  Color _getColor(ZefyrScope editor, ZefyrToolbarTheme theme) {
     if (isAttributeAction) {
       final attribute = kZefyrToolbarAttributeActions[action];
       final isToggled = (attribute is NotusAttribute)
@@ -106,7 +105,7 @@ class ZefyrButton extends StatelessWidget {
   }
 
   VoidCallback _getPressedHandler(
-      ZefyrEditorScope editor, ZefyrToolbarState toolbar) {
+      ZefyrScope editor, ZefyrToolbarState toolbar) {
     if (onPressed != null) {
       return onPressed;
     } else if (isAttributeAction) {
@@ -123,7 +122,7 @@ class ZefyrButton extends StatelessWidget {
     return null;
   }
 
-  void _toggleAttribute(NotusAttribute attribute, ZefyrEditorScope editor) {
+  void _toggleAttribute(NotusAttribute attribute, ZefyrScope editor) {
     final isToggled = editor.selectionStyle.containsSame(attribute);
     if (isToggled) {
       editor.formatSelection(attribute.unset);
@@ -156,7 +155,7 @@ class RawZefyrButton extends StatelessWidget {
     Color iconColor,
     @required this.color,
     @required this.onPressed,
-  })  : child = new Icon(icon, size: size, color: iconColor),
+  })  : child = Icon(icon, size: size, color: iconColor),
         super();
 
   /// Toolbar action associated with this button.
@@ -279,16 +278,20 @@ class _ImageButtonState extends State<ImageButton> {
 
   void _pickFromCamera() async {
     final editor = ZefyrToolbar.of(context).editor;
-    final image = await editor.imageDelegate.pickImage(ImageSource.camera);
-    if (image != null)
+    final image =
+        await editor.imageDelegate.pickImage(editor.imageDelegate.cameraSource);
+    if (image != null) {
       editor.formatSelection(NotusAttribute.embed.image(image));
+    }
   }
 
   void _pickFromGallery() async {
     final editor = ZefyrToolbar.of(context).editor;
-    final image = await editor.imageDelegate.pickImage(ImageSource.gallery);
-    if (image != null)
+    final image = await editor.imageDelegate
+        .pickImage(editor.imageDelegate.gallerySource);
+    if (image != null) {
       editor.formatSelection(NotusAttribute.embed.image(image));
+    }
   }
 }
 
@@ -303,7 +306,7 @@ class _LinkButtonState extends State<LinkButton> {
   final TextEditingController _inputController = TextEditingController();
   Key _inputKey;
   bool _formatError = false;
-  ZefyrEditorScope _editor;
+  ZefyrScope _editor;
 
   bool get isEditing => _inputKey != null;
 
@@ -345,7 +348,7 @@ class _LinkButtonState extends State<LinkButton> {
   void edit() {
     final toolbar = ZefyrToolbar.of(context);
     setState(() {
-      _inputKey = new UniqueKey();
+      _inputKey = UniqueKey();
       _inputController.text = getLink('https://');
       _inputController.addListener(_handleInputChange);
       toolbar.markNeedsRebuild();
@@ -404,7 +407,7 @@ class _LinkButtonState extends State<LinkButton> {
   void copyToClipboard() {
     var link = getLink();
     assert(link != null);
-    Clipboard.setData(new ClipboardData(text: link));
+    Clipboard.setData(ClipboardData(text: link));
   }
 
   void openInBrowser() async {
@@ -479,19 +482,19 @@ class _LinkInput extends StatefulWidget {
   final bool formatError;
 
   const _LinkInput(
-      {Key key, @required this.controller, this.formatError: false})
+      {Key key, @required this.controller, this.formatError = false})
       : super(key: key);
 
   @override
   _LinkInputState createState() {
-    return new _LinkInputState();
+    return _LinkInputState();
   }
 }
 
 class _LinkInputState extends State<_LinkInput> {
   final FocusNode _focusNode = FocusNode();
 
-  ZefyrEditorScope _editor;
+  ZefyrScope _editor;
   bool _didAutoFocus = false;
 
   @override
@@ -505,15 +508,15 @@ class _LinkInputState extends State<_LinkInput> {
     final toolbar = ZefyrToolbar.of(context);
 
     if (_editor != toolbar.editor) {
-      _editor?.setToolbarFocusNode(null);
+      _editor?.toolbarFocusNode = null;
       _editor = toolbar.editor;
-      _editor.setToolbarFocusNode(_focusNode);
+      _editor.toolbarFocusNode = _focusNode;
     }
   }
 
   @override
   void dispose() {
-    _editor?.setToolbarFocusNode(null);
+    _editor?.toolbarFocusNode = null;
     _focusNode.dispose();
     _editor = null;
     super.dispose();
@@ -532,7 +535,7 @@ class _LinkInputState extends State<_LinkInput> {
       focusNode: _focusNode,
       controller: widget.controller,
       autofocus: true,
-      decoration: new InputDecoration(
+      decoration: InputDecoration(
         hintText: 'https://',
         filled: true,
         fillColor: toolbarTheme.color,
@@ -553,7 +556,7 @@ class _LinkView extends StatelessWidget {
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
     final toolbarTheme = ZefyrTheme.of(context).toolbarTheme;
-    Widget widget = new ClipRect(
+    Widget widget = ClipRect(
       child: ListView(
         scrollDirection: Axis.horizontal,
         children: <Widget>[
