@@ -1,8 +1,10 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:math' as math;
 import 'dart:ui' as ui;
 
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:notus/notus.dart';
@@ -10,13 +12,12 @@ import 'package:zefyr/util.dart';
 
 import 'controller.dart';
 import 'editable_box.dart';
-import 'editable_text.dart';
-import 'editor.dart';
+import 'scope.dart';
 
 RenderEditableBox _getEditableBox(HitTestResult result) {
   for (var entry in result.path) {
     if (entry.target is RenderEditableBox) {
-      return entry.target;
+      return entry.target as RenderEditableBox;
     }
   }
   return null;
@@ -24,34 +25,73 @@ RenderEditableBox _getEditableBox(HitTestResult result) {
 
 /// Selection overlay controls selection handles and other gestures.
 class ZefyrSelectionOverlay extends StatefulWidget {
-  const ZefyrSelectionOverlay({
-    Key key,
-    @required this.controller,
-    @required this.controls,
-    @required this.overlay,
-  }) : super(key: key);
+  const ZefyrSelectionOverlay({Key key, @required this.controls})
+      : super(key: key);
 
-  final ZefyrController controller;
   final TextSelectionControls controls;
-  final OverlayState overlay;
 
   @override
-  _ZefyrSelectionOverlayState createState() =>
-      new _ZefyrSelectionOverlayState();
+  _ZefyrSelectionOverlayState createState() => _ZefyrSelectionOverlayState();
 }
 
 class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     implements TextSelectionDelegate {
+  TextSelectionControls _controls;
+  TextSelectionControls get controls => _controls;
+
+  /// Global position of last TapDown event.
+  Offset _lastTapDownPosition;
+
+  /// Global position of last TapDown which is potentially a long press.
+  Offset _longPressPosition;
+
+  OverlayState _overlay;
+  OverlayEntry _toolbar;
+  AnimationController _toolbarController;
+
+  ZefyrScope _scope;
+  ZefyrScope get scope => _scope;
+  TextSelection _selection;
+  FocusOwner _focusOwner;
+
+  bool _didCaretTap = false;
+
+  /// Whether selection controls should be hidden.
+  bool get shouldHideControls {
+    if (!_scope.mode.canSelect) return true;
+    final selection = _scope.selection;
+    final isSelectionCollapsed = selection == null || selection.isCollapsed;
+    if (_scope.mode.canEdit) {
+      return isSelectionCollapsed || _scope.focusOwner != FocusOwner.editor;
+    }
+    return isSelectionCollapsed;
+  }
+
+  void showToolbar() {
+    final toolbarOpacity = _toolbarController.view;
+    _toolbar = OverlayEntry(
+      builder: (context) => FadeTransition(
+        opacity: toolbarOpacity,
+        child: _SelectionToolbar(selectionOverlay: this),
+      ),
+    );
+    _overlay.insert(_toolbar);
+    _toolbarController.forward(from: 0.0);
+  }
+
+  bool get isToolbarVisible => _toolbar != null;
+  bool get isToolbarHidden => _toolbar == null;
+
   @override
   TextEditingValue get textEditingValue =>
-      widget.controller.plainTextEditingValue;
+      _scope.controller.plainTextEditingValue;
 
   set textEditingValue(TextEditingValue value) {
     final cursorPosition = value.selection.extentOffset;
-    final oldText = widget.controller.document.toPlainText();
+    final oldText = _scope.controller.document.toPlainText();
     final newText = value.text;
     final diff = fastDiff(oldText, newText, cursorPosition);
-    widget.controller.replaceText(
+    _scope.controller.replaceText(
         diff.start, diff.deleted.length, diff.inserted,
         selection: value.selection);
   }
@@ -61,75 +101,60 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     // TODO: implement bringIntoView
   }
 
-  bool get isToolbarVisible => _toolbar != null;
-  bool get isToolbarHidden => _toolbar == null;
-
   @override
   void hideToolbar() {
     _didCaretTap = false; // reset double tap.
     _toolbar?.remove();
     _toolbar = null;
-    _toolbarController.stop();
+    _toolbarController?.stop();
   }
 
-  void showToolbar() {
-    final editable = ZefyrEditableText.of(context);
-    assert(editable != null);
-    final toolbarOpacity = _toolbarController.view;
-    _toolbar = new OverlayEntry(
-      builder: (context) => new FadeTransition(
-            opacity: toolbarOpacity,
-            child: new _SelectionToolbar(
-              editable: editable,
-              controls: widget.controls,
-              delegate: this,
-            ),
-          ),
-    );
-    widget.overlay.insert(_toolbar);
-    _toolbarController.forward(from: 0.0);
-  }
-
-  //
-  // Overridden members of State
-  //
+  static const Duration _kFadeDuration = Duration(milliseconds: 150);
 
   @override
   void initState() {
     super.initState();
-    _toolbarController = new AnimationController(
-        duration: _kFadeDuration, vsync: widget.overlay);
+    _controls = widget.controls;
   }
-
-  static const Duration _kFadeDuration = const Duration(milliseconds: 150);
 
   @override
   void didUpdateWidget(ZefyrSelectionOverlay oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.overlay != widget.overlay) {
-      hideToolbar();
-      _toolbarController.dispose();
-      _toolbarController = new AnimationController(
-          duration: _kFadeDuration, vsync: widget.overlay);
-    }
+    _controls = widget.controls;
   }
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final editor = ZefyrEditor.of(context);
-    if (_editor != editor) {
-      _editor?.removeListener(_handleChange);
-      _editor = editor;
-      _editor.addListener(_handleChange);
-      _selection = _editor.selection;
-      _focusOwner = _editor.focusOwner;
+    final scope = ZefyrScope.of(context);
+    if (_scope != scope) {
+      _scope?.removeListener(_handleChange);
+      _scope = scope;
+      _scope.addListener(_handleChange);
+      _selection = _scope.selection;
+      _focusOwner = _scope.focusOwner;
     }
+
+    final overlay = Overlay.of(context, debugRequiredFor: widget);
+    if (_overlay != overlay) {
+      hideToolbar();
+      _overlay = overlay;
+      _toolbarController?.dispose();
+      _toolbarController = null;
+    }
+    if (_toolbarController == null) {
+      _toolbarController = AnimationController(
+        duration: _kFadeDuration,
+        vsync: _overlay,
+      );
+    }
+
+    _toolbar?.markNeedsBuild();
   }
 
   @override
   void dispose() {
-    _editor.removeListener(_handleChange);
+    _scope.removeListener(_handleChange);
     hideToolbar();
     _toolbarController.dispose();
     _toolbarController = null;
@@ -138,50 +163,35 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
 
   @override
   Widget build(BuildContext context) {
-    final overlay = new GestureDetector(
+    final overlay = GestureDetector(
       behavior: HitTestBehavior.translucent,
       onTapDown: _handleTapDown,
       onTap: _handleTap,
       onTapCancel: _handleTapCancel,
       onLongPress: _handleLongPress,
-      child: new Stack(
+      child: Stack(
         fit: StackFit.expand,
         children: <Widget>[
-          new SelectionHandleDriver(
+          SelectionHandleDriver(
             position: _SelectionHandlePosition.base,
-            controls: widget.controls,
+            selectionOverlay: this,
           ),
-          new SelectionHandleDriver(
+          SelectionHandleDriver(
             position: _SelectionHandlePosition.extent,
-            controls: widget.controls,
+            selectionOverlay: this,
           ),
         ],
       ),
     );
-    return new Container(child: overlay);
+    return Container(child: overlay);
   }
 
   //
   // Private members
   //
 
-  /// Global position of last TapDown event.
-  Offset _lastTapDownPosition;
-
-  /// Global position of last TapDown which is potentially a long press.
-  Offset _longPressPosition;
-
-  OverlayEntry _toolbar;
-  AnimationController _toolbarController;
-
-  ZefyrEditorScope _editor;
-  TextSelection _selection;
-  FocusOwner _focusOwner;
-
-  bool _didCaretTap = false;
-
   void _handleChange() {
-    if (_selection != _editor.selection || _focusOwner != _editor.focusOwner) {
+    if (_selection != _scope.selection || _focusOwner != _scope.focusOwner) {
       _updateToolbar();
     }
   }
@@ -190,15 +200,17 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     if (!mounted) {
       return;
     }
-    final editor = ZefyrEditor.of(context);
-    final selection = editor.selection;
-    final focusOwner = editor.focusOwner;
+
+    final selection = _scope.selection;
+    final focusOwner = _scope.focusOwner;
     setState(() {
-      if (focusOwner != FocusOwner.editor) {
+      if (shouldHideControls && isToolbarVisible) {
         hideToolbar();
       } else {
         if (_selection != selection) {
-          if (selection.isCollapsed && isToolbarVisible) hideToolbar();
+          if (selection.isCollapsed && isToolbarVisible) {
+            hideToolbar();
+          }
           _toolbar?.markNeedsBuild();
           if (!selection.isCollapsed && isToolbarHidden) showToolbar();
         } else {
@@ -228,19 +240,18 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     assert(_lastTapDownPosition != null);
     final globalPoint = _lastTapDownPosition;
     _lastTapDownPosition = null;
-    HitTestResult result = new HitTestResult();
+    HitTestResult result = HitTestResult();
     WidgetsBinding.instance.hitTest(result, globalPoint);
 
     RenderEditableProxyBox box = _getEditableBox(result);
     if (box == null) {
-      final editable = ZefyrEditableText.of(context);
-      box = editable.renderContext.closestBoxForGlobalPoint(globalPoint);
+      box = _scope.renderContext.closestBoxForGlobalPoint(globalPoint);
     }
     if (box == null) return null;
 
     final localPoint = box.globalToLocal(globalPoint);
     final position = box.getPositionForOffset(localPoint);
-    final selection = new TextSelection.collapsed(
+    final selection = TextSelection.collapsed(
       offset: position.offset,
       affinity: position.affinity,
     );
@@ -254,13 +265,13 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     } else {
       _didCaretTap = true;
     }
-    widget.controller.updateSelection(selection, source: ChangeSource.local);
+    _scope.controller.updateSelection(selection, source: ChangeSource.local);
   }
 
   void _handleLongPress() {
     final Offset globalPoint = _longPressPosition;
     _longPressPosition = null;
-    HitTestResult result = new HitTestResult();
+    HitTestResult result = HitTestResult();
     WidgetsBinding.instance.hitTest(result, globalPoint);
     final box = _getEditableBox(result);
     if (box == null) {
@@ -269,12 +280,24 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     final localPoint = box.globalToLocal(globalPoint);
     final position = box.getPositionForOffset(localPoint);
     final word = box.getWordBoundary(position);
-    final selection = new TextSelection(
+    final selection = TextSelection(
       baseOffset: word.start,
       extentOffset: word.end,
     );
-    widget.controller.updateSelection(selection, source: ChangeSource.local);
+    _scope.controller.updateSelection(selection, source: ChangeSource.local);
   }
+
+  @override
+  bool get copyEnabled => _scope.mode.canSelect && !_selection.isCollapsed;
+
+  @override
+  bool get cutEnabled => _scope.mode.canEdit && !_selection.isCollapsed;
+
+  @override
+  bool get pasteEnabled => _scope.mode.canEdit;
+
+  @override
+  bool get selectAllEnabled => _scope.mode.canSelect;
 }
 
 enum _SelectionHandlePosition { base, extent }
@@ -283,18 +306,21 @@ class SelectionHandleDriver extends StatefulWidget {
   const SelectionHandleDriver({
     Key key,
     @required this.position,
-    @required this.controls,
-  }) : super(key: key);
+    @required this.selectionOverlay,
+  })  : assert(selectionOverlay != null),
+        super(key: key);
 
   final _SelectionHandlePosition position;
-  final TextSelectionControls controls;
+  final _ZefyrSelectionOverlayState selectionOverlay;
 
   @override
-  _SelectionHandleDriverState createState() =>
-      new _SelectionHandleDriverState();
+  _SelectionHandleDriverState createState() => _SelectionHandleDriverState();
 }
 
-class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
+class _SelectionHandleDriverState extends State<SelectionHandleDriver>
+    with SingleTickerProviderStateMixin {
+  ZefyrScope _scope;
+
   /// Current document selection.
   TextSelection get selection => _selection;
   TextSelection _selection;
@@ -309,26 +335,36 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
   int get documentOffset =>
       isBaseHandle ? selection.baseOffset : selection.extentOffset;
 
-  /// Position in pixels of this selection handle within its paragraph [block].
-  Offset getPosition(RenderEditableBox block) {
+  List<TextSelectionPoint> getEndpointsForSelection(RenderEditableBox block) {
     if (block == null) return null;
 
-    final localSelection = block.getLocalSelection(selection);
-    assert(localSelection != null);
-
-    final boxes = block.getEndpointsForSelection(selection);
-    assert(boxes.isNotEmpty, 'Got empty boxes for selection ${selection}');
-
-    final box = isBaseHandle ? boxes.first : boxes.last;
-    final dx = isBaseHandle ? box.start : box.end;
-    return new Offset(dx, box.bottom);
+    final Offset paintOffset = Offset.zero;
+    final List<ui.TextBox> boxes = block.getEndpointsForSelection(selection);
+    final Offset start =
+        Offset(boxes.first.start, boxes.first.bottom) + paintOffset;
+    final Offset end = Offset(boxes.last.end, boxes.last.bottom) + paintOffset;
+    return <TextSelectionPoint>[
+      TextSelectionPoint(start, boxes.first.direction),
+      TextSelectionPoint(end, boxes.last.direction),
+    ];
   }
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final editable = ZefyrEditableText.of(context);
-    _selection = editable.selection;
+    final scope = ZefyrScope.of(context);
+    if (_scope != scope) {
+      _scope?.removeListener(_handleScopeChange);
+      _scope = scope;
+      _scope.addListener(_handleScopeChange);
+    }
+    _selection = _scope.selection;
+  }
+
+  @override
+  void dispose() {
+    _scope?.removeListener(_handleScopeChange);
+    super.dispose();
   }
 
   //
@@ -337,48 +373,102 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
 
   @override
   Widget build(BuildContext context) {
-    final editor = ZefyrEditor.of(context);
-    final editable = ZefyrEditableText.of(context);
-    if (selection == null ||
-        selection.isCollapsed ||
-        widget.controls == null ||
-        editor.focusOwner != FocusOwner.editor) {
-      return new Container();
+    if (widget.selectionOverlay.shouldHideControls) {
+      return Container();
     }
-    final block = editable.renderContext.boxForTextOffset(documentOffset);
-    final position = getPosition(block);
-    Widget handle;
-    if (position == null) {
-      handle = new Container();
-    } else {
-      final handleType = isBaseHandle
-          ? TextSelectionHandleType.left
-          : TextSelectionHandleType.right;
-      handle = new Positioned(
-        left: position.dx,
-        top: position.dy,
-        child: widget.controls.buildHandle(
-          context,
-          handleType,
-          block.preferredLineHeight,
-        ),
-      );
-      handle = new CompositedTransformFollower(
-        link: block.layerLink,
-        showWhenUnlinked: false,
-        child: new Stack(
-          overflow: Overflow.visible,
-          children: <Widget>[handle],
-        ),
-      );
+    final block = _scope.renderContext.boxForTextOffset(documentOffset);
+    if (block == null) {
+      // TODO: For some reason sometimes we get updates when render boxes
+      //      are in process of rebuilding so we don't have access to them here.
+      //      As a workaround we just return empty container. There is usually
+      //      another rebuild right after which "fixes" the view.
+      //      Example: when toolbar button is toggled changing style of current
+      //      selection.
+      return Container();
     }
-    // Always return this gesture detector even if handle is an empty container
-    // This way we prevent drag gesture from being canceled in case current
-    // position is somewhere outside of any visible paragraph block.
-    return new GestureDetector(
-      onPanStart: _handleDragStart,
-      onPanUpdate: _handleDragUpdate,
-      child: handle,
+
+    final List<TextSelectionPoint> endpoints = getEndpointsForSelection(block);
+    Offset point;
+    TextSelectionHandleType type;
+
+    switch (widget.position) {
+      case _SelectionHandlePosition.base:
+        point = endpoints[0].point;
+        type = _chooseType(endpoints[0], TextSelectionHandleType.left,
+            TextSelectionHandleType.right);
+        break;
+      case _SelectionHandlePosition.extent:
+        // [endpoints] will only contain 1 point for collapsed selections, in
+        // which case we shouldn't be building the [end] handle.
+        assert(endpoints.length == 2);
+        point = endpoints[1].point;
+        type = _chooseType(endpoints[1], TextSelectionHandleType.right,
+            TextSelectionHandleType.left);
+        break;
+    }
+
+    final Size viewport = block.size;
+    point = Offset(
+      point.dx.clamp(0.0, viewport.width),
+      point.dy.clamp(0.0, viewport.height),
+    );
+
+    final Offset handleAnchor =
+        widget.selectionOverlay.controls.getHandleAnchor(
+      type,
+      block.preferredLineHeight,
+    );
+    final Size handleSize = widget.selectionOverlay.controls.getHandleSize(
+      block.preferredLineHeight,
+    );
+    final Rect handleRect = Rect.fromLTWH(
+      // Put handleAnchor on top of point
+      point.dx - handleAnchor.dx,
+      point.dy - handleAnchor.dy,
+      handleSize.width,
+      handleSize.height,
+    );
+
+    // Make sure the GestureDetector is big enough to be easily interactive.
+    final Rect interactiveRect = handleRect.expandToInclude(
+      Rect.fromCircle(
+          center: handleRect.center, radius: kMinInteractiveDimension / 2),
+    );
+    final RelativeRect padding = RelativeRect.fromLTRB(
+      math.max((interactiveRect.width - handleRect.width) / 2, 0),
+      math.max((interactiveRect.height - handleRect.height) / 2, 0),
+      math.max((interactiveRect.width - handleRect.width) / 2, 0),
+      math.max((interactiveRect.height - handleRect.height) / 2, 0),
+    );
+
+    return CompositedTransformFollower(
+      link: block.layerLink,
+      offset: interactiveRect.topLeft,
+      showWhenUnlinked: false,
+      child: Container(
+        alignment: Alignment.topLeft,
+        width: interactiveRect.width,
+        height: interactiveRect.height,
+        child: GestureDetector(
+          behavior: HitTestBehavior.translucent,
+          dragStartBehavior: DragStartBehavior.start,
+          onPanStart: _handleDragStart,
+          onPanUpdate: _handleDragUpdate,
+          child: Padding(
+            padding: EdgeInsets.only(
+              left: padding.left,
+              top: padding.top,
+              right: padding.right,
+              bottom: padding.bottom,
+            ),
+            child: widget.selectionOverlay.controls.buildHandle(
+              context,
+              type,
+              block.preferredLineHeight,
+            ),
+          ),
+        ),
+      ),
     );
   }
 
@@ -386,7 +476,32 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
   // Private members
   //
 
+  TextSelectionHandleType _chooseType(
+    TextSelectionPoint endpoint,
+    TextSelectionHandleType ltrType,
+    TextSelectionHandleType rtlType,
+  ) {
+    if (selection.isCollapsed) return TextSelectionHandleType.collapsed;
+
+    assert(endpoint.direction != null);
+    switch (endpoint.direction) {
+      case TextDirection.ltr:
+        return ltrType;
+      case TextDirection.rtl:
+        return rtlType;
+    }
+    return null;
+  }
+
   Offset _dragPosition;
+
+  void _handleScopeChange() {
+    if (_selection != _scope.selection) {
+      setState(() {
+        _selection = _scope.selection;
+      });
+    }
+  }
 
   void _handleDragStart(DragStartDetails details) {
     _dragPosition = details.globalPosition;
@@ -395,9 +510,7 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
   void _handleDragUpdate(DragUpdateDetails details) {
     _dragPosition += details.delta;
     final globalPoint = _dragPosition;
-    final editor = ZefyrEditor.of(context);
-    final editable = ZefyrEditableText.of(context);
-    final paragraph = editable.renderContext.boxForGlobalPoint(globalPoint);
+    final paragraph = _scope.renderContext.boxForGlobalPoint(globalPoint);
     if (paragraph == null) {
       return;
     }
@@ -414,7 +527,7 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
     }
 
     if (newSelection != _selection) {
-      editor.updateSelection(newSelection, source: ChangeSource.local);
+      _scope.updateSelection(newSelection, source: ChangeSource.local);
     }
   }
 }
@@ -422,22 +535,20 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
 class _SelectionToolbar extends StatefulWidget {
   const _SelectionToolbar({
     Key key,
-    @required this.editable,
-    @required this.controls,
-    @required this.delegate,
+    @required this.selectionOverlay,
   }) : super(key: key);
 
-  final ZefyrEditableTextScope editable;
-  final TextSelectionControls controls;
-  final TextSelectionDelegate delegate;
+  final _ZefyrSelectionOverlayState selectionOverlay;
 
   @override
-  _SelectionToolbarState createState() => new _SelectionToolbarState();
+  _SelectionToolbarState createState() => _SelectionToolbarState();
 }
 
 class _SelectionToolbarState extends State<_SelectionToolbar> {
-  ZefyrEditableTextScope get editable => widget.editable;
-  TextSelection get selection => widget.delegate.textEditingValue.selection;
+  TextSelectionControls get controls => widget.selectionOverlay.controls;
+  ZefyrScope get scope => widget.selectionOverlay.scope;
+  TextSelection get selection =>
+      widget.selectionOverlay.textEditingValue.selection;
 
   @override
   Widget build(BuildContext context) {
@@ -446,29 +557,48 @@ class _SelectionToolbarState extends State<_SelectionToolbar> {
 
   Widget _buildToolbar(BuildContext context) {
     final base = selection.baseOffset;
-    // TODO: Editable is not refreshed and may contain stale renderContext instance.
-    final block = editable.renderContext.boxForTextOffset(base);
+    final block = scope.renderContext.boxForTextOffset(base);
     if (block == null) {
       return Container();
     }
     final boxes = block.getEndpointsForSelection(selection);
-
     // Find the horizontal midpoint, just above the selected text.
-    final Offset midpoint = new Offset(
+    Offset midpoint = Offset(
       (boxes.length == 1)
           ? (boxes[0].start + boxes[0].end) / 2.0
           : (boxes[0].start + boxes[1].start) / 2.0,
       boxes[0].bottom - block.preferredLineHeight,
     );
+    List<TextSelectionPoint> endpoints;
+    if (boxes.length == 1) {
+      midpoint = Offset((boxes[0].start + boxes[0].end) / 2.0,
+          boxes[0].bottom - block.preferredLineHeight);
+      final Offset start = Offset(boxes[0].start, block.preferredLineHeight);
+      endpoints = <TextSelectionPoint>[TextSelectionPoint(start, null)];
+    } else {
+      midpoint = Offset((boxes[0].start + boxes[1].start) / 2.0,
+          boxes[0].bottom - block.preferredLineHeight);
+      final Offset start = Offset(boxes.first.start, boxes.first.bottom);
+      final Offset end = Offset(boxes.last.end, boxes.last.bottom);
+      endpoints = <TextSelectionPoint>[
+        TextSelectionPoint(start, boxes.first.direction),
+        TextSelectionPoint(end, boxes.last.direction),
+      ];
+    }
 
-    final Rect editingRegion = new Rect.fromPoints(
+    final Rect editingRegion = Rect.fromPoints(
       block.localToGlobal(Offset.zero),
       block.localToGlobal(block.size.bottomRight(Offset.zero)),
     );
 
-    final toolbar = widget.controls
-        .buildToolbar(context, editingRegion, midpoint, widget.delegate);
-    return new CompositedTransformFollower(
+    final toolbar = controls.buildToolbar(
+        context,
+        editingRegion,
+        block.preferredLineHeight,
+        midpoint,
+        endpoints,
+        widget.selectionOverlay);
+    return CompositedTransformFollower(
       link: block.layerLink,
       showWhenUnlinked: false,
       offset: -editingRegion.topLeft,
