@@ -12,10 +12,10 @@ class NotusMarkdownCodec extends Codec<Delta, String> {
 
   @override
   Converter<String, Delta> get decoder =>
-      throw new UnimplementedError('Decoding is not implemented yet.');
+      throw UnimplementedError('Decoding is not implemented yet.');
 
   @override
-  Converter<Delta, String> get encoder => new _NotusMarkdownEncoder();
+  Converter<Delta, String> get encoder => _NotusMarkdownEncoder();
 }
 
 class _NotusMarkdownEncoder extends Converter<Delta, String> {
@@ -29,11 +29,11 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
 
   @override
   String convert(Delta input) {
-    final iterator = new DeltaIterator(input);
-    final buffer = new StringBuffer();
-    final lineBuffer = new StringBuffer();
+    final iterator = DeltaIterator(input);
+    final buffer = StringBuffer();
+    final lineBuffer = StringBuffer();
     NotusAttribute<String> currentBlockStyle;
-    NotusStyle currentInlineStyle = new NotusStyle();
+    NotusStyle currentInlineStyle = NotusStyle();
     List<String> currentBlockLines = [];
 
     void _handleBlock(NotusAttribute<String> blockStyle) {
@@ -86,7 +86,7 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
       if (lf == -1) {
         _handleSpan(op.data, op.attributes);
       } else {
-        StringBuffer span = new StringBuffer();
+        StringBuffer span = StringBuffer();
         for (var i = 0; i < op.data.length; i++) {
           if (op.data.codeUnitAt(i) == 0x0A) {
             if (span.isNotEmpty) {
@@ -112,9 +112,9 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
   }
 
   String _writeLine(String text, NotusStyle style) {
-    StringBuffer buffer = new StringBuffer();
+    StringBuffer buffer = StringBuffer();
     if (style.contains(NotusAttribute.heading)) {
-      _writeAttribute(buffer, style.get(NotusAttribute.heading));
+      _writeAttribute(buffer, style.get<int>(NotusAttribute.heading));
     }
 
     // Write the text itself
@@ -157,19 +157,19 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
   }
 
   void _writeAttribute(StringBuffer buffer, NotusAttribute attribute,
-      {bool close: false}) {
+      {bool close = false}) {
     if (attribute == NotusAttribute.bold) {
       _writeBoldTag(buffer);
     } else if (attribute == NotusAttribute.italic) {
       _writeItalicTag(buffer);
     } else if (attribute.key == NotusAttribute.link.key) {
-      _writeLinkTag(buffer, attribute, close: close);
+      _writeLinkTag(buffer, attribute as NotusAttribute<String>, close: close);
     } else if (attribute.key == NotusAttribute.heading.key) {
-      _writeHeadingTag(buffer, attribute);
+      _writeHeadingTag(buffer, attribute as NotusAttribute<int>);
     } else if (attribute.key == NotusAttribute.block.key) {
-      _writeBlockTag(buffer, attribute, close: close);
+      _writeBlockTag(buffer, attribute as NotusAttribute<String>, close: close);
     } else {
-      throw new ArgumentError('Cannot handle $attribute');
+      throw ArgumentError('Cannot handle $attribute');
     }
   }
 
@@ -182,7 +182,7 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
   }
 
   void _writeLinkTag(StringBuffer buffer, NotusAttribute<String> link,
-      {bool close: false}) {
+      {bool close = false}) {
     if (close) {
       buffer.write('](${link.value})');
     } else {
@@ -196,7 +196,7 @@ class _NotusMarkdownEncoder extends Converter<Delta, String> {
   }
 
   void _writeBlockTag(StringBuffer buffer, NotusAttribute<String> block,
-      {bool close: false}) {
+      {bool close = false}) {
     if (block == NotusAttribute.code) {
       if (close) {
         buffer.write('\n```');
