@@ -40,11 +40,11 @@ class NotusDocument {
   /// Creates new empty Notus document.
   NotusDocument()
       : _heuristics = NotusHeuristics.fallback,
-        _delta = new Delta()..insert('\n') {
+        _delta = Delta()..insert('\n') {
     _loadDocument(_delta);
   }
 
-  NotusDocument.fromJson(dynamic data)
+  NotusDocument.fromJson(List data)
       : _heuristics = NotusHeuristics.fallback,
         _delta = Delta.fromJson(data) {
     _loadDocument(_delta);
@@ -61,7 +61,7 @@ class NotusDocument {
 
   /// The root node of this document tree.
   RootNode get root => _root;
-  final RootNode _root = new RootNode();
+  final RootNode _root = RootNode();
 
   /// Length of this document.
   int get length => _root.length;
@@ -70,10 +70,10 @@ class NotusDocument {
   Stream<NotusChange> get changes => _controller.stream;
 
   final StreamController<NotusChange> _controller =
-      new StreamController.broadcast();
+      StreamController.broadcast();
 
   /// Returns contents of this document as [Delta].
-  Delta toDelta() => new Delta.from(_delta);
+  Delta toDelta() => Delta.from(_delta);
   Delta _delta;
 
   /// Returns plain text representation of this document.
@@ -104,7 +104,7 @@ class NotusDocument {
     assert(index >= 0);
     assert(text.isNotEmpty);
     text = _sanitizeString(text);
-    if (text.isEmpty) return new Delta();
+    if (text.isEmpty) return Delta();
     final change = _heuristics.applyInsertRules(this, index, text);
     compose(change, ChangeSource.local);
     return change;
@@ -136,7 +136,7 @@ class NotusDocument {
   Delta replace(int index, int length, String text) {
     assert(index >= 0 && (text.isNotEmpty || length > 0),
         'With index $index, length $length and text "$text"');
-    Delta delta = new Delta();
+    Delta delta = Delta();
     // We have to compose before applying delete rules
     // Otherwise delete would be operating on stale document snapshot.
     if (text.isNotEmpty) {
@@ -161,10 +161,24 @@ class NotusDocument {
   /// unchanged and no [NotusChange] is published to [changes] stream.
   Delta format(int index, int length, NotusAttribute attribute) {
     assert(index >= 0 && length >= 0 && attribute != null);
-    final change = _heuristics.applyFormatRules(this, index, length, attribute);
-    if (change.isNotEmpty) {
-      compose(change, ChangeSource.local);
+
+    Delta change = Delta();
+
+    if (attribute is EmbedAttribute && length > 0) {
+      // Must delete selected length of text before applying embed attribute
+      // since inserting an embed in non-empty selection is essentially a
+      // replace operation.
+      change = delete(index, length);
+      length = 0;
     }
+
+    final formatChange =
+        _heuristics.applyFormatRules(this, index, length, attribute);
+    if (formatChange.isNotEmpty) {
+      compose(formatChange, ChangeSource.local);
+      change = change.compose(formatChange);
+    }
+
     return change;
   }
 
@@ -225,10 +239,10 @@ class NotusDocument {
     _delta = _delta.compose(change);
 
     if (_delta != _root.toDelta()) {
-      throw new StateError('Compose produced inconsistent results. '
+      throw StateError('Compose produced inconsistent results. '
           'This is likely due to a bug in the library. Tried to compose change $change from $source.');
     }
-    _controller.add(new NotusChange(before, change, source));
+    _controller.add(NotusChange(before, change, source));
   }
 
   //
@@ -265,7 +279,7 @@ class NotusDocument {
       if (op.isInsert) {
         _root.insert(offset, op.data, style);
       } else {
-        throw new ArgumentError.value(doc,
+        throw ArgumentError.value(doc,
             "Document Delta can only contain insert operations but ${op.key} found.");
       }
       offset += op.length;
