@@ -40,9 +40,10 @@ abstract class NotusAttributeBuilder<T> implements NotusAttributeKey<T> {
 
   final String key;
   final NotusAttributeScope scope;
-  NotusAttribute<T> get unset => new NotusAttribute<T>._(key, scope, null);
+
+  NotusAttribute<T> get unset => NotusAttribute<T>._(key, scope, null);
   NotusAttribute<T> withValue(T value) =>
-      new NotusAttribute<T>._(key, scope, value);
+      NotusAttribute<T>._(key, scope, value);
 }
 
 /// Style attribute applicable to a segment of a Notus document.
@@ -82,18 +83,20 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   // Inline attributes
 
   /// Bold style attribute.
-  static const bold = const _BoldAttribute();
+  static const bold = _BoldAttribute();
 
   /// Italic style attribute.
-  static const italic = const _ItalicAttribute();
+  static const italic = _ItalicAttribute();
 
   /// Link style attribute.
-  static const link = const LinkAttributeBuilder._();
+  // ignore: const_eval_throws_exception
+  static const link = LinkAttributeBuilder._();
 
   // Line attributes
 
   /// Heading style attribute.
-  static const heading = const HeadingAttributeBuilder._();
+  // ignore: const_eval_throws_exception
+  static const heading = HeadingAttributeBuilder._();
 
   /// Alias for [NotusAttribute.heading.level1].
   static NotusAttribute<int> get h1 => heading.level1;
@@ -105,7 +108,8 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   static NotusAttribute<int> get h3 => heading.level3;
 
   /// Block attribute
-  static const block = const BlockAttributeBuilder._();
+  // ignore: const_eval_throws_exception
+  static const block = BlockAttributeBuilder._();
 
   /// Alias for [NotusAttribute.block.bulletList].
   static NotusAttribute<String> get ul => block.bulletList;
@@ -120,12 +124,14 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   static NotusAttribute<String> get code => block.code;
 
   /// Embed style attribute.
-  static const embed = const EmbedAttributeBuilder._();
+  // ignore: const_eval_throws_exception
+  static const embed = EmbedAttributeBuilder._();
 
-  factory NotusAttribute._fromKeyValue(String key, T value) {
-    if (!_registry.containsKey(key))
-      throw new ArgumentError.value(
+  static NotusAttribute _fromKeyValue(String key, dynamic value) {
+    if (!_registry.containsKey(key)) {
+      throw ArgumentError.value(
           key, 'No attribute with key "$key" registered.');
+    }
     final builder = _registry[key];
     return builder.withValue(value);
   }
@@ -154,7 +160,7 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   ///
   /// When composed into a rich text document, unset attributes remove
   /// associated style.
-  NotusAttribute<T> get unset => new NotusAttribute<T>._(key, scope, null);
+  NotusAttribute<T> get unset => NotusAttribute<T>._(key, scope, null);
 
   /// Returns `true` if this attribute is an unset attribute.
   bool get isUnset => value == null;
@@ -163,10 +169,10 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   bool get isInline => scope == NotusAttributeScope.inline;
 
   NotusAttribute<T> withValue(T value) =>
-      new NotusAttribute<T>._(key, scope, value);
+      NotusAttribute<T>._(key, scope, value);
 
   @override
-  bool operator ==(other) {
+  bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other is! NotusAttribute<T>) return false;
     NotusAttribute<T> typedOther = other;
@@ -190,17 +196,17 @@ class NotusStyle {
 
   final Map<String, NotusAttribute> _data;
 
-  static NotusStyle fromJson(Map data) {
-    if (data == null) return new NotusStyle();
+  static NotusStyle fromJson(Map<String, dynamic> data) {
+    if (data == null) return NotusStyle();
 
-    final result = data.map((key, value) {
-      var attr = new NotusAttribute._fromKeyValue(key, value);
-      return new MapEntry<String, NotusAttribute>(key, attr);
+    final result = data.map((String key, dynamic value) {
+      var attr = NotusAttribute._fromKeyValue(key, value);
+      return MapEntry<String, NotusAttribute>(key, attr);
     });
-    return new NotusStyle._(result);
+    return NotusStyle._(result);
   }
 
-  NotusStyle() : _data = new Map<String, NotusAttribute>();
+  NotusStyle() : _data = Map<String, NotusAttribute>();
 
   /// Returns `true` if this attribute set is empty.
   bool get isEmpty => _data.isEmpty;
@@ -228,7 +234,7 @@ class NotusStyle {
   /// [attribute].
   bool containsSame(NotusAttribute attribute) {
     assert(attribute != null);
-    return get(attribute) == attribute;
+    return get<dynamic>(attribute) == attribute;
   }
 
   /// Returns value of specified attribute [key] in this set.
@@ -246,9 +252,9 @@ class NotusStyle {
 
   /// Puts [attribute] into this attribute set and returns result as a new set.
   NotusStyle put(NotusAttribute attribute) {
-    final result = new Map<String, NotusAttribute>.from(_data);
+    final result = Map<String, NotusAttribute>.from(_data);
     result[attribute.key] = attribute;
-    return new NotusStyle._(result);
+    return NotusStyle._(result);
   }
 
   /// Merges this attribute set with [attribute] and returns result as a new
@@ -260,19 +266,19 @@ class NotusStyle {
   /// See also [put] method which does not perform compaction and allows
   /// constructing styles with "unset" values.
   NotusStyle merge(NotusAttribute attribute) {
-    final merged = new Map<String, NotusAttribute>.from(_data);
+    final merged = Map<String, NotusAttribute>.from(_data);
     if (attribute.isUnset) {
       merged.remove(attribute.key);
     } else {
       merged[attribute.key] = attribute;
     }
-    return new NotusStyle._(merged);
+    return NotusStyle._(merged);
   }
 
   /// Merges all attributes from [other] into this style and returns result
   /// as a new instance of [NotusStyle].
   NotusStyle mergeAll(NotusStyle other) {
-    var result = new NotusStyle._(_data);
+    var result = NotusStyle._(_data);
     for (var value in other.values) {
       result = result.merge(value);
     }
@@ -282,19 +288,19 @@ class NotusStyle {
   /// Removes [attributes] from this style and returns new instance of
   /// [NotusStyle] containing result.
   NotusStyle removeAll(Iterable<NotusAttribute> attributes) {
-    final merged = new Map<String, NotusAttribute>.from(_data);
+    final merged = Map<String, NotusAttribute>.from(_data);
     attributes.map((item) => item.key).forEach(merged.remove);
-    return new NotusStyle._(merged);
+    return NotusStyle._(merged);
   }
 
   /// Returns JSON-serializable representation of this style.
   Map<String, dynamic> toJson() => _data.isEmpty
       ? null
-      : _data.map(
-          (_, value) => new MapEntry<String, dynamic>(value.key, value.value));
+      : _data.map<String, dynamic>((String _, NotusAttribute value) =>
+          MapEntry<String, dynamic>(value.key, value.value));
 
   @override
-  bool operator ==(other) {
+  bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other is! NotusStyle) return false;
     NotusStyle typedOther = other;
@@ -328,11 +334,12 @@ class _ItalicAttribute extends NotusAttribute<bool> {
 /// [NotusAttribute.link] instead.
 class LinkAttributeBuilder extends NotusAttributeBuilder<String> {
   static const _kLink = 'a';
+
   const LinkAttributeBuilder._() : super._(_kLink, NotusAttributeScope.inline);
 
   /// Creates a link attribute with specified link [value].
   NotusAttribute<String> fromString(String value) =>
-      new NotusAttribute<String>._(key, scope, value);
+      NotusAttribute<String>._(key, scope, value);
 }
 
 /// Builder for heading attribute styles.
@@ -341,17 +348,18 @@ class LinkAttributeBuilder extends NotusAttributeBuilder<String> {
 /// [NotusAttribute.heading] instead.
 class HeadingAttributeBuilder extends NotusAttributeBuilder<int> {
   static const _kHeading = 'heading';
+
   const HeadingAttributeBuilder._()
       : super._(_kHeading, NotusAttributeScope.line);
 
   /// Level 1 heading, equivalent of `H1` in HTML.
-  NotusAttribute<int> get level1 => new NotusAttribute<int>._(key, scope, 1);
+  NotusAttribute<int> get level1 => NotusAttribute<int>._(key, scope, 1);
 
   /// Level 2 heading, equivalent of `H2` in HTML.
-  NotusAttribute<int> get level2 => new NotusAttribute<int>._(key, scope, 2);
+  NotusAttribute<int> get level2 => NotusAttribute<int>._(key, scope, 2);
 
   /// Level 3 heading, equivalent of `H3` in HTML.
-  NotusAttribute<int> get level3 => new NotusAttribute<int>._(key, scope, 3);
+  NotusAttribute<int> get level3 => NotusAttribute<int>._(key, scope, 3);
 }
 
 /// Builder for block attribute styles (number/bullet lists, code and quote).
@@ -360,23 +368,24 @@ class HeadingAttributeBuilder extends NotusAttributeBuilder<int> {
 /// [NotusAttribute.block] instead.
 class BlockAttributeBuilder extends NotusAttributeBuilder<String> {
   static const _kBlock = 'block';
+
   const BlockAttributeBuilder._() : super._(_kBlock, NotusAttributeScope.line);
 
   /// Formats a block of lines as a bullet list.
   NotusAttribute<String> get bulletList =>
-      new NotusAttribute<String>._(key, scope, 'ul');
+      NotusAttribute<String>._(key, scope, 'ul');
 
   /// Formats a block of lines as a number list.
   NotusAttribute<String> get numberList =>
-      new NotusAttribute<String>._(key, scope, 'ol');
+      NotusAttribute<String>._(key, scope, 'ol');
 
   /// Formats a block of lines as a code snippet, using monospace font.
   NotusAttribute<String> get code =>
-      new NotusAttribute<String>._(key, scope, 'code');
+      NotusAttribute<String>._(key, scope, 'code');
 
   /// Formats a block of lines as a quote.
   NotusAttribute<String> get quote =>
-      new NotusAttribute<String>._(key, scope, 'quote');
+      NotusAttribute<String>._(key, scope, 'quote');
 }
 
 class EmbedAttributeBuilder
@@ -401,7 +410,7 @@ class EmbedAttributeBuilder
 enum EmbedType { horizontalRule, image }
 
 class EmbedAttribute extends NotusAttribute<Map<String, dynamic>> {
-  static const _kValueEquality = const MapEquality<String, dynamic>();
+  static const _kValueEquality = MapEquality<String, dynamic>();
   static const _kEmbed = 'embed';
   static const _kHorizontalRuleEmbed = 'hr';
   static const _kImageEmbed = 'image';
@@ -424,7 +433,7 @@ class EmbedAttribute extends NotusAttribute<Map<String, dynamic>> {
   }
 
   @override
-  NotusAttribute<Map<String, dynamic>> get unset => new EmbedAttribute._(null);
+  NotusAttribute<Map<String, dynamic>> get unset => EmbedAttribute._(null);
 
   @override
   bool operator ==(other) {
