@@ -21,7 +21,7 @@ class CatchAllInsertRule extends InsertRule {
 
   @override
   Delta apply(Delta document, int index, String text) {
-    return new Delta()
+    return Delta()
       ..retain(index)
       ..insert(text);
   }
@@ -43,12 +43,12 @@ class PreserveLineStyleOnSplitRule extends InsertRule {
   Delta apply(Delta document, int index, String text) {
     if (text != '\n') return null;
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     final before = iter.skip(index);
     final after = iter.next();
     if (isEdgeLineSplit(before, after)) return null;
-    Delta result = new Delta()..retain(index);
-    if (after.data.indexOf('\n') >= 0) {
+    Delta result = Delta()..retain(index);
+    if (after.data.contains('\n')) {
       // It is not allowed to combine line and inline styles in insert
       // operation containing line-break together with other characters.
       // The only scenario we get such operation is when the text is plain.
@@ -82,17 +82,17 @@ class ResetLineFormatOnNewLineRule extends InsertRule {
   Delta apply(Delta document, int index, String text) {
     if (text != '\n') return null;
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     iter.skip(index);
     final target = iter.next();
 
     if (target.data.startsWith('\n')) {
-      Map<String, dynamic> resetStyle = null;
+      Map<String, dynamic> resetStyle;
       if (target.attributes != null &&
           target.attributes.containsKey(NotusAttribute.heading.key)) {
         resetStyle = NotusAttribute.heading.unset.toJson();
       }
-      return new Delta()
+      return Delta()
         ..retain(index)
         ..insert('\n', target.attributes)
         ..retain(1, resetStyle)
@@ -117,7 +117,7 @@ class AutoExitBlockRule extends InsertRule {
   Delta apply(Delta document, int index, String text) {
     if (text != '\n') return null;
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     final previous = iter.skip(index);
     final target = iter.next();
     final isInBlock = target.isNotPlain &&
@@ -129,7 +129,7 @@ class AutoExitBlockRule extends InsertRule {
       var attributes =
           target.attributes != null ? target.attributes : <String, dynamic>{};
       attributes.addAll(NotusAttribute.block.unset.toJson());
-      return new Delta()..retain(index)..retain(1, attributes);
+      return Delta()..retain(index)..retain(1, attributes);
     }
     return null;
   }
@@ -144,7 +144,7 @@ class PreserveInlineStylesRule extends InsertRule {
     // This rule is only applicable to characters other than line-break.
     if (text.contains('\n')) return null;
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     final previous = iter.skip(index);
     // If there is a line-break in previous chunk, there should be no inline
     // styles. Also if there is no previous operation we are at the beginning
@@ -155,7 +155,7 @@ class PreserveInlineStylesRule extends InsertRule {
     final hasLink =
         (attributes != null && attributes.containsKey(NotusAttribute.link.key));
     if (!hasLink) {
-      return new Delta()
+      return Delta()
         ..retain(index)
         ..insert(text, attributes);
     }
@@ -164,7 +164,7 @@ class PreserveInlineStylesRule extends InsertRule {
     // Link style should NOT be preserved on the boundaries.
     var noLinkAttributes = previous.attributes;
     noLinkAttributes.remove(NotusAttribute.link.key);
-    final noLinkResult = new Delta()
+    final noLinkResult = Delta()
       ..retain(index)
       ..insert(text, noLinkAttributes.isEmpty ? null : noLinkAttributes);
     final next = iter.next();
@@ -180,7 +180,7 @@ class PreserveInlineStylesRule extends InsertRule {
     // We must make sure links are identical in previous and next operations.
     if (attributes[NotusAttribute.link.key] ==
         nextAttributes[NotusAttribute.link.key]) {
-      return new Delta()
+      return Delta()
         ..retain(index)
         ..insert(text, attributes);
     } else {
@@ -200,7 +200,7 @@ class AutoFormatLinksRule extends InsertRule {
     // everything else.
     if (text != ' ') return null;
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     final previous = iter.skip(index);
     // No previous operation means no link.
     if (previous == null) return null;
@@ -220,7 +220,7 @@ class AutoFormatLinksRule extends InsertRule {
 
       attributes
           .addAll(NotusAttribute.link.fromString(link.toString()).toJson());
-      return new Delta()
+      return Delta()
         ..retain(index - candidate.length)
         ..retain(candidate.length, attributes)
         ..insert(text, previous.attributes);
@@ -239,13 +239,13 @@ class ForceNewlineForInsertsAroundEmbedRule extends InsertRule {
 
   @override
   Delta apply(Delta document, int index, String text) {
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     final previous = iter.skip(index);
     final target = iter.next();
     final beforeEmbed = target.data == EmbedNode.kPlainTextPlaceholder;
     final afterEmbed = previous?.data == EmbedNode.kPlainTextPlaceholder;
     if (beforeEmbed || afterEmbed) {
-      final delta = new Delta()..retain(index);
+      final delta = Delta()..retain(index);
       if (beforeEmbed && !text.endsWith('\n')) {
         return delta..insert(text)..insert('\n');
       }
@@ -276,7 +276,7 @@ class PreserveBlockStyleOnPasteRule extends InsertRule {
       return null;
     }
 
-    DeltaIterator iter = new DeltaIterator(document);
+    DeltaIterator iter = DeltaIterator(document);
     iter.skip(index);
 
     // Look for next line-break.
@@ -290,8 +290,8 @@ class PreserveBlockStyleOnPasteRule extends InsertRule {
       }
     }
 
-    Map<String, dynamic> resetStyle = null;
-    Map<String, dynamic> blockStyle = null;
+    Map<String, dynamic> resetStyle;
+    Map<String, dynamic> blockStyle;
     if (lineStyle != null) {
       if (lineStyle.containsKey(NotusAttribute.heading.key)) {
         resetStyle = NotusAttribute.heading.unset.toJson();
@@ -305,7 +305,7 @@ class PreserveBlockStyleOnPasteRule extends InsertRule {
     }
 
     final lines = text.split('\n');
-    Delta result = new Delta()..retain(index);
+    Delta result = Delta()..retain(index);
     for (int i = 0; i < lines.length; i++) {
       final line = lines[i];
       if (line.isNotEmpty) {
