@@ -19,14 +19,14 @@ void main() {
 
   group('$NotusMarkdownCodec.encode', () {
     test('split adjacent paragraphs', () {
-      final delta = new Delta()..insert('First line\nSecond line\n');
+      final delta = Delta()..insert('First line\nSecond line\n');
       final result = notusMarkdown.encode(delta);
       expect(result, 'First line\n\nSecond line\n\n');
     });
 
     test('bold italic', () {
       runFor(NotusAttribute<bool> attribute, String expected) {
-        final delta = new Delta()
+        final delta = Delta()
           ..insert('This ')
           ..insert('house', attribute.toJson())
           ..insert(' is a ')
@@ -44,10 +44,10 @@ void main() {
     test('intersecting inline styles', () {
       final b = NotusAttribute.bold.toJson();
       final i = NotusAttribute.italic.toJson();
-      final bi = new Map<String, dynamic>.from(b);
+      final bi = Map<String, dynamic>.from(b);
       bi.addAll(i);
 
-      final delta = new Delta()
+      final delta = Delta()
         ..insert('This ')
         ..insert('house', b)
         ..insert(' is a ', bi)
@@ -61,7 +61,7 @@ void main() {
     test('normalize inline styles', () {
       final b = NotusAttribute.bold.toJson();
       final i = NotusAttribute.italic.toJson();
-      final delta = new Delta()
+      final delta = Delta()
         ..insert('This')
         ..insert(' house ', b)
         ..insert('is a')
@@ -75,7 +75,7 @@ void main() {
     test('links', () {
       final b = NotusAttribute.bold.toJson();
       final link = NotusAttribute.link.fromString('https://github.com');
-      final delta = new Delta()
+      final delta = Delta()
         ..insert('This')
         ..insert(' house ', b)
         ..insert('is a')
@@ -88,9 +88,7 @@ void main() {
 
     test('heading styles', () {
       runFor(NotusAttribute<int> attribute, String source, String expected) {
-        final delta = new Delta()
-          ..insert(source)
-          ..insert('\n', attribute.toJson());
+        final delta = Delta()..insert(source)..insert('\n', attribute.toJson());
         final result = notusMarkdown.encode(delta);
         expect(result, expected);
       }
@@ -102,9 +100,7 @@ void main() {
 
     test('block styles', () {
       runFor(NotusAttribute<String> attribute, String source, String expected) {
-        final delta = new Delta()
-          ..insert(source)
-          ..insert('\n', attribute.toJson());
+        final delta = Delta()..insert(source)..insert('\n', attribute.toJson());
         final result = notusMarkdown.encode(delta);
         expect(result, expected);
       }
@@ -117,7 +113,7 @@ void main() {
 
     test('multiline blocks', () {
       runFor(NotusAttribute<String> attribute, String source, String expected) {
-        final delta = new Delta()
+        final delta = Delta()
           ..insert(source)
           ..insert('\n', attribute.toJson())
           ..insert(source)
@@ -141,7 +137,7 @@ void main() {
 
 final doc =
     r'[{"insert":"Zefyr"},{"insert":"\n","attributes":{"heading":1}},{"insert":"Soft and gentle rich text editing for Flutter applications.","attributes":{"i":true}},{"insert":"\nZefyr is an "},{"insert":"early preview","attributes":{"b":true}},{"insert":" open source library.\nDocumentation"},{"insert":"\n","attributes":{"heading":3}},{"insert":"Quick Start"},{"insert":"\n","attributes":{"block":"ul"}},{"insert":"Data format and Document Model"},{"insert":"\n","attributes":{"block":"ul"}},{"insert":"Style attributes"},{"insert":"\n","attributes":{"block":"ul"}},{"insert":"Heuristic rules"},{"insert":"\n","attributes":{"block":"ul"}},{"insert":"Clean and modern look"},{"insert":"\n","attributes":{"heading":2}},{"insert":"Zefyr’s rich text editor is built with simplicity and flexibility in mind. It provides clean interface for distraction-free editing. Think Medium.com-like experience.\nimport ‘package:flutter/material.dart’;"},{"insert":"\n","attributes":{"block":"code"}},{"insert":"import ‘package:notus/notus.dart’;"},{"insert":"\n\n","attributes":{"block":"code"}},{"insert":"void main() {"},{"insert":"\n","attributes":{"block":"code"}},{"insert":" print(“Hello world!”);"},{"insert":"\n","attributes":{"block":"code"}},{"insert":"}"},{"insert":"\n","attributes":{"block":"code"}}]';
-final delta = Delta.fromJson(json.decode(doc));
+final delta = Delta.fromJson(json.decode(doc) as List);
 
 final expectedMarkdown = '''
 # Zefyr
