@@ -9,18 +9,18 @@ import 'package:notus/notus.dart';
 import 'matchers.dart';
 
 NotusDocument dartconfDoc() {
-  Delta delta = new Delta()..insert('DartConf\nLos Angeles\n');
-  return new NotusDocument.fromDelta(delta);
+  Delta delta = Delta()..insert('DartConf\nLos Angeles\n');
+  return NotusDocument.fromDelta(delta);
 }
 
 NotusDocument dartconfEmbedDoc() {
   final hr = NotusAttribute.embed.horizontalRule.toJson();
-  Delta delta = new Delta()
+  Delta delta = Delta()
     ..insert('DartConf\n')
     ..insert(kZeroWidthSpace, hr)
     ..insert('\n')
     ..insert('Los Angeles\n');
-  return new NotusDocument.fromDelta(delta);
+  return NotusDocument.fromDelta(delta);
 }
 
 final ul = NotusAttribute.ul.toJson();
@@ -29,24 +29,24 @@ final h1 = NotusAttribute.h1.toJson();
 void main() {
   group('$NotusDocument', () {
     test('validates for doc delta', () {
-      var badDelta = new Delta()
+      var badDelta = Delta()
         ..insert('Text')
         ..retain(5)
         ..insert('\n');
       expect(() {
-        new NotusDocument.fromDelta(badDelta);
+        NotusDocument.fromDelta(badDelta);
       }, throwsArgumentError);
     });
 
     test('empty document contains single empty line', () {
-      NotusDocument doc = new NotusDocument();
+      NotusDocument doc = NotusDocument();
       expect(doc.toPlainText(), '\n');
     });
 
     test('json serialization', () {
       final original = dartconfDoc();
       final jsonData = json.encode(original);
-      final doc = NotusDocument.fromJson(json.decode(jsonData));
+      final doc = NotusDocument.fromJson(json.decode(jsonData) as List);
       expect(doc.toDelta(), original.toDelta());
       expect(json.encode(doc), jsonData);
     });
@@ -67,9 +67,9 @@ void main() {
     });
 
     test('document delta must end with line-break character', () {
-      Delta delta = new Delta()..insert('DartConf\nLos Angeles');
+      Delta delta = Delta()..insert('DartConf\nLos Angeles');
       expect(() {
-        new NotusDocument.fromDelta(delta);
+        NotusDocument.fromDelta(delta);
       }, throwsA(const TypeMatcher<AssertionError>()));
     });
 
@@ -105,7 +105,7 @@ void main() {
     test('format returns actual change delta', () {
       NotusDocument doc = dartconfDoc();
       final change = doc.format(0, 15, NotusAttribute.ul);
-      final expectedChange = new Delta()
+      final expectedChange = Delta()
         ..retain(8)
         ..retain(1, ul)
         ..retain(11)
@@ -116,7 +116,7 @@ void main() {
     test('format updates document delta', () {
       NotusDocument doc = dartconfDoc();
       doc.format(0, 15, NotusAttribute.ul);
-      final expectedDoc = new Delta()
+      final expectedDoc = Delta()
         ..insert('DartConf')
         ..insert('\n', ul)
         ..insert('Los Angeles')
@@ -127,7 +127,7 @@ void main() {
     test('format allows zero-length updates', () {
       NotusDocument doc = dartconfDoc();
       doc.format(0, 0, NotusAttribute.ul);
-      final expectedDoc = new Delta()
+      final expectedDoc = Delta()
         ..insert('DartConf')
         ..insert('\n', ul)
         ..insert('Los Angeles')
@@ -147,7 +147,7 @@ void main() {
       NotusDocument doc = dartconfDoc();
       doc.format(0, 15, NotusAttribute.ul);
       final change = doc.insert(8, '\n');
-      final expectedChange = new Delta()
+      final expectedChange = Delta()
         ..retain(8)
         ..insert('\n', ul);
       expect(change, expectedChange);
@@ -157,7 +157,7 @@ void main() {
       NotusDocument doc = dartconfDoc();
       doc.format(0, 15, NotusAttribute.ul);
       doc.insert(8, '\n');
-      final expectedDoc = new Delta()
+      final expectedDoc = Delta()
         ..insert('DartConf')
         ..insert('\n\n', ul)
         ..insert('Los Angeles')
@@ -182,7 +182,7 @@ void main() {
     test('compose throws assert error if change is empty', () {
       NotusDocument doc = dartconfDoc();
       expect(() {
-        doc.compose(new Delta()..retain(1), ChangeSource.local);
+        doc.compose(Delta()..retain(1), ChangeSource.local);
       }, throwsA(const TypeMatcher<AssertionError>()));
     });
 
@@ -219,7 +219,7 @@ void main() {
       doc.close();
       expect(doc.isClosed, isTrue);
       expect(() {
-        doc.compose(new Delta()..insert('a'), ChangeSource.local);
+        doc.compose(Delta()..insert('a'), ChangeSource.local);
       }, throwsAssertionError);
       expect(() {
         doc.insert(0, 'a');
@@ -247,7 +247,7 @@ void main() {
       LineNode line = doc.root.children.elementAt(1);
       EmbedNode embed = line.first;
       expect(embed.toPlainText(), EmbedNode.kPlainTextPlaceholder);
-      final style = new NotusStyle().merge(NotusAttribute.embed.horizontalRule);
+      final style = NotusStyle().merge(NotusAttribute.embed.horizontalRule);
       expect(embed.style, style);
     });
 
@@ -260,7 +260,7 @@ void main() {
       LineNode line = doc.root.children.elementAt(1);
       EmbedNode embed = line.first;
       expect(embed.toPlainText(), EmbedNode.kPlainTextPlaceholder);
-      final style = new NotusStyle().merge(NotusAttribute.embed.horizontalRule);
+      final style = NotusStyle().merge(NotusAttribute.embed.horizontalRule);
       expect(embed.style, style);
     });
 
@@ -276,7 +276,7 @@ void main() {
       LineNode line = doc.root.children.elementAt(1);
       EmbedNode embed = line.first;
       expect(embed.toPlainText(), EmbedNode.kPlainTextPlaceholder);
-      final style = new NotusStyle().merge(NotusAttribute.embed.horizontalRule);
+      final style = NotusStyle().merge(NotusAttribute.embed.horizontalRule);
       expect(embed.style, style);
     });
 
@@ -322,6 +322,18 @@ void main() {
     test('replace text with embed', () {
       final doc = dartconfDoc();
       doc.format(4, 4, NotusAttribute.embed.horizontalRule);
+      expect(doc.root.children, hasLength(3));
+      expect(doc.root.children.elementAt(0).toPlainText(), 'Dart\n');
+      expect(doc.root.children.elementAt(1).toPlainText(),
+          '${EmbedNode.kPlainTextPlaceholder}\n');
+      expect(doc.root.children.elementAt(2).toPlainText(), 'Los Angeles\n');
+    });
+
+    test('replace embed with embed', () {
+      final doc = dartconfDoc();
+      doc.format(4, 4, NotusAttribute.embed.horizontalRule);
+      doc.format(5, 1, NotusAttribute.embed.horizontalRule);
+
       expect(doc.root.children, hasLength(3));
       expect(doc.root.children.elementAt(0).toPlainText(), 'Dart\n');
       expect(doc.root.children.elementAt(1).toPlainText(),
