@@ -10,15 +10,15 @@ final bold = NotusAttribute.bold.toJson();
 
 void main() {
   group('$PreserveLineStyleOnMergeRule', () {
-    final rule = new PreserveLineStyleOnMergeRule();
+    final rule = PreserveLineStyleOnMergeRule();
     test('preserves block style', () {
       final ul = NotusAttribute.ul.toJson();
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Title\nOne')
         ..insert('\n', ul)
         ..insert('Two\n');
       final actual = rule.apply(doc, 9, 1);
-      final expected = new Delta()
+      final expected = Delta()
         ..retain(9)
         ..delete(1)
         ..retain(3)
@@ -28,12 +28,12 @@ void main() {
 
     test('resets block style', () {
       final unsetUl = NotusAttribute.block.unset.toJson();
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Title\nOne')
         ..insert('\n', NotusAttribute.ul.toJson())
         ..insert('Two\n');
       final actual = rule.apply(doc, 5, 1);
-      final expected = new Delta()
+      final expected = Delta()
         ..retain(5)
         ..delete(1)
         ..retain(3)
@@ -43,12 +43,12 @@ void main() {
   });
 
   group('$CatchAllDeleteRule', () {
-    final rule = new CatchAllDeleteRule();
+    final rule = CatchAllDeleteRule();
 
     test('applies change as-is', () {
-      final doc = new Delta()..insert('Document\n');
+      final doc = Delta()..insert('Document\n');
       final actual = rule.apply(doc, 3, 5);
-      final expected = new Delta()
+      final expected = Delta()
         ..retain(3)
         ..delete(5);
       expect(actual, expected);
@@ -56,40 +56,40 @@ void main() {
   });
 
   group('$EnsureEmbedLineRule', () {
-    final rule = new EnsureEmbedLineRule();
+    final rule = EnsureEmbedLineRule();
 
     test('ensures line-break before embed', () {
       final hr = NotusAttribute.embed.horizontalRule;
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Document\n')
         ..insert(kZeroWidthSpace, hr.toJson())
         ..insert('\n');
       final actual = rule.apply(doc, 8, 1);
-      final expected = new Delta()..retain(8);
+      final expected = Delta()..retain(8);
       expect(actual, expected);
     });
 
     test('ensures line-break after embed', () {
       final hr = NotusAttribute.embed.horizontalRule;
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Document\n')
         ..insert(kZeroWidthSpace, hr.toJson())
         ..insert('\n');
       final actual = rule.apply(doc, 10, 1);
-      final expected = new Delta()..retain(11);
+      final expected = Delta()..retain(11);
       expect(actual, expected);
     });
 
     test('still deletes everything between embeds', () {
       final hr = NotusAttribute.embed.horizontalRule;
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Document\n')
         ..insert(kZeroWidthSpace, hr.toJson())
         ..insert('\nSome text\n')
         ..insert(kZeroWidthSpace, hr.toJson())
         ..insert('\n');
       final actual = rule.apply(doc, 10, 11);
-      final expected = new Delta()
+      final expected = Delta()
         ..retain(11)
         ..delete(9);
       expect(actual, expected);
@@ -97,7 +97,7 @@ void main() {
 
     test('allows deleting empty line after embed', () {
       final hr = NotusAttribute.embed.horizontalRule;
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Document\n')
         ..insert(kZeroWidthSpace, hr.toJson())
         ..insert('\n')
@@ -105,7 +105,7 @@ void main() {
         ..insert('Text')
         ..insert('\n');
       final actual = rule.apply(doc, 10, 1);
-      final expected = new Delta()
+      final expected = Delta()
         ..retain(11)
         ..delete(1);
       expect(actual, expected);
@@ -113,7 +113,7 @@ void main() {
 
     test('allows deleting empty line(s) before embed', () {
       final hr = NotusAttribute.embed.horizontalRule;
-      final doc = new Delta()
+      final doc = Delta()
         ..insert('Document\n')
         ..insert('\n')
         ..insert('\n')
