@@ -8,7 +8,7 @@ import 'package:notus/notus.dart';
 import 'package:quill_delta/quill_delta.dart';
 import 'package:zefyr/util.dart';
 
-const TextSelection _kZeroSelection = const TextSelection.collapsed(
+const TextSelection _kZeroSelection = TextSelection.collapsed(
   offset: 0,
   affinity: TextAffinity.upstream,
 );
@@ -44,18 +44,24 @@ class ZefyrController extends ChangeNotifier {
   /// Source of the last text or selection change.
   ChangeSource get lastChangeSource => _lastChangeSource;
 
+  /// Store any styles attribute that got toggled by the tap of a button
+  /// and that has not been applied yet.
+  /// It gets reseted after each format action within the [document].
+  NotusStyle get toggledStyles => _toggledStyles;
+  NotusStyle _toggledStyles = NotusStyle();
+
   /// Updates selection with specified [value].
   ///
   /// [value] and [source] cannot be `null`.
   void updateSelection(TextSelection value,
-      {ChangeSource source: ChangeSource.remote}) {
+      {ChangeSource source = ChangeSource.remote}) {
     _updateSelectionSilent(value, source: source);
     notifyListeners();
   }
 
   // Updates selection without triggering notifications to listeners.
   void _updateSelectionSilent(TextSelection value,
-      {ChangeSource source: ChangeSource.remote}) {
+      {ChangeSource source = ChangeSource.remote}) {
     assert(value != null && source != null);
     _selection = value;
     _lastChangeSource = source;
@@ -76,7 +82,7 @@ class ZefyrController extends ChangeNotifier {
   ///
   /// If composing this change fails then this method throws [ComposeError].
   void compose(Delta change,
-      {TextSelection selection, ChangeSource source: ChangeSource.remote}) {
+      {TextSelection selection, ChangeSource source = ChangeSource.remote}) {
     if (change.isNotEmpty) {
       _document.compose(change, source);
     }
@@ -98,13 +104,38 @@ class ZefyrController extends ChangeNotifier {
     notifyListeners();
   }
 
+  /// Replaces [length] characters in the document starting at [index] with
+  /// provided [text].
+  ///
+  /// Resulting change is registered as produced by user action, e.g.
+  /// using [ChangeSource.local].
+  ///
+  /// It also applies the toggledStyle if needed. And then it resets it
+  /// in any cases as we don't want to keep it except on inserts.
+  ///
+  /// Optionally updates selection if provided.
   void replaceText(int index, int length, String text,
       {TextSelection selection}) {
     Delta delta;
 
     if (length > 0 || text.isNotEmpty) {
       delta = document.replace(index, length, text);
+      // If the delta is a classical insert operation and we have toggled
+      // some style, then we apply it to our document.
+      if (delta != null &&
+          toggledStyles.isNotEmpty &&
+          delta.length == 2 &&
+          delta[1].isInsert) {
+        // Apply it.
+        Delta retainDelta = Delta()
+          ..retain(index)
+          ..retain(1, toggledStyles.toJson());
+        document.compose(retainDelta, ChangeSource.local);
+      }
     }
+
+    // Always reset it after any user action, even if it has not been applied.
+    _toggledStyles = NotusStyle();
 
     if (selection != null) {
       if (delta == null) {
@@ -112,7 +143,7 @@ class ZefyrController extends ChangeNotifier {
       } else {
         // need to transform selection position in case actual delta
         // is different from user's version (in deletes and inserts).
-        Delta user = new Delta()
+        Delta user = Delta()
           ..retain(index)
           ..insert(text)
           ..delete(length);
@@ -133,12 +164,19 @@ class ZefyrController extends ChangeNotifier {
   void formatText(int index, int length, NotusAttribute attribute) {
     final change = document.format(index, length, attribute);
     _lastChangeSource = ChangeSource.local;
+
+    if (length == 0 &&
+        (attribute.key == NotusAttribute.bold.key ||
+            attribute.key == NotusAttribute.italic.key)) {
+      // Add the attribute to our toggledStyle. It will be used later upon insertion.
+      _toggledStyles = toggledStyles.put(attribute);
+    }
+
     // Transform selection against the composed change and give priority to
     // the change. This is needed in cases when format operation actually
     // inserts data into the document (e.g. embeds).
     final base = change.transformPosition(_selection.baseOffset);
-    final extent =
-        change.transformPosition(_selection.extentOffset);
+    final extent = change.transformPosition(_selection.extentOffset);
     final adjustedSelection =
         _selection.copyWith(baseOffset: base, extentOffset: extent);
     if (_selection != adjustedSelection) {
@@ -154,17 +192,25 @@ class ZefyrController extends ChangeNotifier {
     formatText(index, length, attribute);
   }
 
+  /// Returns style of specified text range.
+  ///
+  /// If nothing is selected but we've toggled an attribute,
+  ///  we also merge those in our style before returning.
   NotusStyle getSelectionStyle() {
     int start = _selection.start;
     int length = _selection.end - start;
-    return _document.collectStyle(start, length);
+    var lineStyle = _document.collectStyle(start, length);
+
+    lineStyle = lineStyle.mergeAll(toggledStyles);
+
+    return lineStyle;
   }
 
   TextEditingValue get plainTextEditingValue {
-    return new TextEditingValue(
+    return TextEditingValue(
       text: document.toPlainText(),
       selection: selection,
-      composing: new TextRange.collapsed(0),
+      composing: TextRange.collapsed(0),
     );
   }
 
