@@ -1,29 +1,31 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-import 'dart:async';
-
-import 'package:collection/collection.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:notus/notus.dart';
 
 import 'code.dart';
 import 'common.dart';
 import 'controller.dart';
-import 'editable_box.dart';
+import 'cursor_timer.dart';
 import 'editor.dart';
 import 'image.dart';
 import 'input.dart';
 import 'list.dart';
+import 'mode.dart';
 import 'paragraph.dart';
 import 'quote.dart';
 import 'render_context.dart';
+import 'scope.dart';
 import 'selection.dart';
+import 'theme.dart';
 
 /// Core widget responsible for editing Zefyr documents.
 ///
-/// Depends on presence of [ZefyrTheme] somewhere up the widget tree.
+/// Depends on presence of [ZefyrTheme] and [ZefyrScope] somewhere up the
+/// widget tree.
 ///
 /// Consider using [ZefyrEditor] which wraps this widget and adds a toolbar to
 /// edit style attributes.
@@ -33,59 +35,48 @@ class ZefyrEditableText extends StatefulWidget {
     @required this.controller,
     @required this.focusNode,
     @required this.imageDelegate,
-    this.autofocus: true,
-    this.enabled: true,
-    this.padding: const EdgeInsets.symmetric(horizontal: 16.0),
+    this.selectionControls,
+    this.autofocus = true,
+    this.mode = ZefyrMode.edit,
+    this.padding = const EdgeInsets.symmetric(horizontal: 16.0),
     this.physics,
-  }) : super(key: key);
+  })  : assert(mode != null),
+        assert(controller != null),
+        assert(focusNode != null),
+        super(key: key);
 
+  /// Controls the document being edited.
   final ZefyrController controller;
+
+  /// Controls whether this editor has keyboard focus.
   final FocusNode focusNode;
   final ZefyrImageDelegate imageDelegate;
+
+  /// Whether this text field should focus itself if nothing else is already
+  /// focused.
+  ///
+  /// If true, the keyboard will open as soon as this text field obtains focus.
+  /// Otherwise, the keyboard is only shown after the user taps the text field.
+  ///
+  /// Defaults to true. Cannot be null.
   final bool autofocus;
-  final bool enabled;
+
+  /// Editing mode of this text field.
+  final ZefyrMode mode;
+
+  /// Controls physics of scrollable text field.
   final ScrollPhysics physics;
+
+  /// Optional delegate for building the text selection handles and toolbar.
+  ///
+  /// If not provided then platform-specific implementation is used by default.
+  final TextSelectionControls selectionControls;
 
   /// Padding around editable area.
   final EdgeInsets padding;
 
-  static ZefyrEditableTextScope of(BuildContext context) {
-    final ZefyrEditableTextScope result =
-        context.inheritFromWidgetOfExactType(ZefyrEditableTextScope);
-    return result;
-  }
-
   @override
-  _ZefyrEditableTextState createState() => new _ZefyrEditableTextState();
-}
-
-/// Provides access to shared state of [ZefyrEditableText].
-class ZefyrEditableTextScope extends InheritedWidget {
-  static const _kEquality = const SetEquality<RenderEditableBox>();
-
-  ZefyrEditableTextScope({
-    Key key,
-    @required Widget child,
-    @required this.selection,
-    @required this.showCursor,
-    @required this.renderContext,
-    @required this.imageDelegate,
-  })  : _activeBoxes = new Set.from(renderContext.active),
-        super(key: key, child: child);
-
-  final TextSelection selection;
-  final ValueNotifier<bool> showCursor;
-  final ZefyrRenderContext renderContext;
-  final ZefyrImageDelegate imageDelegate;
-  final Set<RenderEditableBox> _activeBoxes;
-
-  @override
-  bool updateShouldNotify(ZefyrEditableTextScope oldWidget) {
-    return selection != oldWidget.selection ||
-        showCursor != oldWidget.showCursor ||
-        imageDelegate != oldWidget.imageDelegate ||
-        !_kEquality.equals(_activeBoxes, oldWidget._activeBoxes);
-  }
+  _ZefyrEditableTextState createState() => _ZefyrEditableTextState();
 }
 
 class _ZefyrEditableTextState extends State<ZefyrEditableText>
@@ -94,16 +85,14 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
   // New public members
   //
 
-  /// Focus node of this widget.
-  FocusNode get focusNode => widget.focusNode;
-
   /// Document controlled by this widget.
   NotusDocument get document => widget.controller.document;
 
   /// Current text selection.
   TextSelection get selection => widget.controller.selection;
-  ZefyrRenderContext get renderContext => _renderContext;
-  ValueNotifier<bool> get showCursor => _cursorTimer.value;
+
+  FocusNode _focusNode;
+  FocusAttachment _focusAttachment;
 
   /// Express interest in interacting with the keyboard.
   ///
@@ -113,10 +102,30 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
   /// focus, the control will then attach to the keyboard and request that the
   /// keyboard become visible.
   void requestKeyboard() {
-    if (focusNode.hasFocus)
+    if (_focusNode.hasFocus) {
       _input.openConnection(widget.controller.plainTextEditingValue);
-    else
-      FocusScope.of(context).requestFocus(focusNode);
+    } else {
+      FocusScope.of(context).requestFocus(_focusNode);
+    }
+  }
+
+  void focusOrUnfocusIfNeeded() {
+    if (!_didAutoFocus && widget.autofocus && widget.mode.canEdit) {
+      FocusScope.of(context).autofocus(_focusNode);
+      _didAutoFocus = true;
+    }
+    if (!widget.mode.canEdit && _focusNode.hasFocus) {
+      _didAutoFocus = false;
+      _focusNode.unfocus();
+    }
+  }
+
+  TextSelectionControls defaultSelectionControls(BuildContext context) {
+    TargetPlatform platform = Theme.of(context).platform;
+    if (platform == TargetPlatform.iOS) {
+      return cupertinoTextSelectionControls;
+    }
+    return materialTextSelectionControls;
   }
 
   //
@@ -125,62 +134,69 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
 
   @override
   Widget build(BuildContext context) {
-    FocusScope.of(context).reparentIfNeeded(focusNode);
+    _focusAttachment.reparent();
     super.build(context); // See AutomaticKeepAliveState.
-    ZefyrEditor.of(context);
 
     Widget body = ListBody(children: _buildChildren(context));
     if (widget.padding != null) {
-      body = new Padding(padding: widget.padding, child: body);
+      body = Padding(padding: widget.padding, child: body);
     }
-    final scrollable = SingleChildScrollView(
+
+    body = SingleChildScrollView(
       physics: widget.physics,
       controller: _scrollController,
       child: body,
     );
 
-    final overlay = Overlay.of(context, debugRequiredFor: widget);
-    final layers = <Widget>[scrollable];
-    if (widget.enabled) {
-      layers.add(ZefyrSelectionOverlay(
-        controller: widget.controller,
-        controls: cupertinoTextSelectionControls,
-        overlay: overlay,
-      ));
-    }
+    final layers = <Widget>[body];
+    layers.add(ZefyrSelectionOverlay(
+      controls: widget.selectionControls ?? defaultSelectionControls(context),
+    ));
 
-    return new ZefyrEditableTextScope(
-      selection: selection,
-      showCursor: showCursor,
-      renderContext: renderContext,
-      imageDelegate: widget.imageDelegate,
-      child: Stack(fit: StackFit.expand, children: layers),
-    );
+    return Stack(fit: StackFit.expand, children: layers);
   }
 
   @override
   void initState() {
+    _focusNode = widget.focusNode;
     super.initState();
-    _input = new InputConnectionController(_handleRemoteValueChange);
+    _focusAttachment = _focusNode.attach(context);
+    _input = InputConnectionController(_handleRemoteValueChange);
     _updateSubscriptions();
   }
 
   @override
   void didUpdateWidget(ZefyrEditableText oldWidget) {
     super.didUpdateWidget(oldWidget);
+    if (_focusNode != widget.focusNode) {
+      _focusAttachment.detach();
+      _focusNode = widget.focusNode;
+      _focusAttachment = _focusNode.attach(context);
+    }
     _updateSubscriptions(oldWidget);
-    if (!_didAutoFocus && widget.autofocus && widget.enabled) {
-      FocusScope.of(context).autofocus(focusNode);
-      _didAutoFocus = true;
+    focusOrUnfocusIfNeeded();
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    final scope = ZefyrScope.of(context);
+    if (_renderContext != scope.renderContext) {
+      _renderContext?.removeListener(_handleRenderContextChange);
+      _renderContext = scope.renderContext;
+      _renderContext.addListener(_handleRenderContextChange);
     }
-    if (!widget.enabled && focusNode.hasFocus) {
-      _didAutoFocus = false;
-      focusNode.unfocus();
+    if (_cursorTimer != scope.cursorTimer) {
+      _cursorTimer?.stop();
+      _cursorTimer = scope.cursorTimer;
+      _cursorTimer.startOrStop(_focusNode, selection);
     }
+    focusOrUnfocusIfNeeded();
   }
 
   @override
   void dispose() {
+    _focusAttachment.detach();
     _cancelSubscriptions();
     super.dispose();
   }
@@ -190,15 +206,15 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
   //
 
   @override
-  bool get wantKeepAlive => focusNode.hasFocus;
+  bool get wantKeepAlive => _focusNode.hasFocus;
 
   //
   // Private members
   //
 
-  final ScrollController _scrollController = new ScrollController();
-  final ZefyrRenderContext _renderContext = new ZefyrRenderContext();
-  final _CursorTimer _cursorTimer = new _CursorTimer();
+  final ScrollController _scrollController = ScrollController();
+  ZefyrRenderContext _renderContext;
+  CursorTimer _cursorTimer;
   InputConnectionController _input;
   bool _didAutoFocus = false;
 
@@ -213,33 +229,32 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
   Widget _defaultChildBuilder(BuildContext context, Node node) {
     if (node is LineNode) {
       if (node.hasEmbed) {
-        return new RawZefyrLine(node: node);
+        return RawZefyrLine(node: node);
       } else if (node.style.contains(NotusAttribute.heading)) {
-        return new ZefyrHeading(node: node);
+        return ZefyrHeading(node: node);
       }
-      return new ZefyrParagraph(node: node);
+      return ZefyrParagraph(node: node);
     }
 
     final BlockNode block = node;
     final blockStyle = block.style.get(NotusAttribute.block);
     if (blockStyle == NotusAttribute.block.code) {
-      return new ZefyrCode(node: node);
+      return ZefyrCode(node: block);
     } else if (blockStyle == NotusAttribute.block.bulletList) {
-      return new ZefyrList(node: node);
+      return ZefyrList(node: block);
     } else if (blockStyle == NotusAttribute.block.numberList) {
-      return new ZefyrList(node: node);
+      return ZefyrList(node: block);
     } else if (blockStyle == NotusAttribute.block.quote) {
-      return new ZefyrQuote(node: node);
+      return ZefyrQuote(node: block);
     }
 
-    throw new UnimplementedError('Block format $blockStyle.');
+    throw UnimplementedError('Block format $blockStyle.');
   }
 
   void _updateSubscriptions([ZefyrEditableText oldWidget]) {
     if (oldWidget == null) {
-      _renderContext.addListener(_handleRenderContextChange);
       widget.controller.addListener(_handleLocalValueChange);
-      focusNode.addListener(_handleFocusChange);
+      _focusNode.addListener(_handleFocusChange);
       return;
     }
 
@@ -257,22 +272,21 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
 
   void _cancelSubscriptions() {
     _renderContext.removeListener(_handleRenderContextChange);
-    _renderContext.dispose();
     widget.controller.removeListener(_handleLocalValueChange);
-    focusNode.removeListener(_handleFocusChange);
+    _focusNode.removeListener(_handleFocusChange);
     _input.closeConnection();
     _cursorTimer.stop();
   }
 
   // Triggered for both text and selection changes.
   void _handleLocalValueChange() {
-    if (widget.enabled &&
+    if (widget.mode.canEdit &&
         widget.controller.lastChangeSource == ChangeSource.local) {
       // Only request keyboard for user actions.
       requestKeyboard();
     }
     _input.updateRemoteValue(widget.controller.plainTextEditingValue);
-    _cursorTimer.startOrStop(focusNode, selection);
+    _cursorTimer.startOrStop(_focusNode, selection);
     setState(() {
       // nothing to update internally.
     });
@@ -280,8 +294,8 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
 
   void _handleFocusChange() {
     _input.openOrCloseConnection(
-        focusNode, widget.controller.plainTextEditingValue);
-    _cursorTimer.startOrStop(focusNode, selection);
+        _focusNode, widget.controller.plainTextEditingValue);
+    _cursorTimer.startOrStop(_focusNode, selection);
     updateKeepAlive();
   }
 
@@ -295,44 +309,5 @@ class _ZefyrEditableTextState extends State<ZefyrEditableText>
     setState(() {
       // nothing to update internally.
     });
-  }
-}
-
-/// Helper class that keeps state relevant to the cursor.
-class _CursorTimer {
-  static const _kCursorBlinkHalfPeriod = const Duration(milliseconds: 500);
-
-  Timer _timer;
-  final ValueNotifier<bool> _showCursor = new ValueNotifier<bool>(false);
-
-  ValueNotifier<bool> get value => _showCursor;
-
-  void _cursorTick(Timer timer) {
-    _showCursor.value = !_showCursor.value;
-  }
-
-  /// Starts cursor timer.
-  void start() {
-    _showCursor.value = true;
-    _timer = new Timer.periodic(_kCursorBlinkHalfPeriod, _cursorTick);
-  }
-
-  /// Stops cursor timer.
-  void stop() {
-    _timer?.cancel();
-    _timer = null;
-    _showCursor.value = false;
-  }
-
-  /// Starts or stops cursor timer based on current state of [focusNode]
-  /// and [selection].
-  void startOrStop(FocusNode focusNode, TextSelection selection) {
-    final hasFocus = focusNode.hasFocus;
-    final selectionCollapsed = selection.isCollapsed;
-    if (_timer == null && hasFocus && selectionCollapsed) {
-      start();
-    } else if (_timer != null && (!hasFocus || !selectionCollapsed)) {
-      stop();
-    }
   }
 }
