@@ -1,162 +1,18 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:notus/notus.dart';
 
 import 'controller.dart';
 import 'editable_text.dart';
 import 'image.dart';
+import 'mode.dart';
 import 'scaffold.dart';
+import 'scope.dart';
 import 'theme.dart';
 import 'toolbar.dart';
-
-class ZefyrEditorScope extends ChangeNotifier {
-  ZefyrEditorScope({
-    @required ZefyrImageDelegate imageDelegate,
-    @required ZefyrController controller,
-    @required FocusNode focusNode,
-    @required FocusScopeNode focusScope,
-  })  : _controller = controller,
-        _imageDelegate = imageDelegate,
-        _focusScope = focusScope,
-        _focusNode = focusNode {
-    _selectionStyle = _controller.getSelectionStyle();
-    _selection = _controller.selection;
-    _controller.addListener(_handleControllerChange);
-    _focusNode.addListener(_handleFocusChange);
-  }
-
-  bool _disposed = false;
-
-  ZefyrImageDelegate _imageDelegate;
-  ZefyrImageDelegate get imageDelegate => _imageDelegate;
-
-  FocusScopeNode _focusScope;
-  FocusNode _focusNode;
-
-  ZefyrController _controller;
-  NotusStyle get selectionStyle => _selectionStyle;
-  NotusStyle _selectionStyle;
-  TextSelection get selection => _selection;
-  TextSelection _selection;
-
-  @override
-  void dispose() {
-    assert(!_disposed);
-    _controller.removeListener(_handleControllerChange);
-    _focusNode.removeListener(_handleFocusChange);
-    _disposed = true;
-    super.dispose();
-  }
-
-  void _updateControllerIfNeeded(ZefyrController value) {
-    if (_controller != value) {
-      _controller.removeListener(_handleControllerChange);
-      _controller = value;
-      _selectionStyle = _controller.getSelectionStyle();
-      _selection = _controller.selection;
-      _controller.addListener(_handleControllerChange);
-      notifyListeners();
-    }
-  }
-
-  void _updateFocusNodeIfNeeded(FocusNode value) {
-    if (_focusNode != value) {
-      _focusNode.removeListener(_handleFocusChange);
-      _focusNode = value;
-      _focusNode.addListener(_handleFocusChange);
-      notifyListeners();
-    }
-  }
-
-  void _updateImageDelegateIfNeeded(ZefyrImageDelegate value) {
-    if (_imageDelegate != value) {
-      _imageDelegate = value;
-      notifyListeners();
-    }
-  }
-
-  void _handleControllerChange() {
-    assert(!_disposed);
-    final attrs = _controller.getSelectionStyle();
-    final selection = _controller.selection;
-    if (_selectionStyle != attrs || _selection != selection) {
-      _selectionStyle = attrs;
-      _selection = _controller.selection;
-      notifyListeners();
-    }
-  }
-
-  void _handleFocusChange() {
-    assert(!_disposed);
-    if (focusOwner == FocusOwner.none && !_selection.isCollapsed) {
-      // Collapse selection if there is nothing focused.
-      _controller.updateSelection(_selection.copyWith(
-        baseOffset: _selection.extentOffset,
-        extentOffset: _selection.extentOffset,
-      ));
-    }
-    notifyListeners();
-  }
-
-  FocusNode _toolbarFocusNode;
-
-  void setToolbarFocusNode(FocusNode node) {
-    assert(!_disposed || node == null);
-    if (_toolbarFocusNode != node) {
-      _toolbarFocusNode?.removeListener(_handleFocusChange);
-      _toolbarFocusNode = node;
-      _toolbarFocusNode?.addListener(_handleFocusChange);
-      // We do not notify listeners here because it will happen when
-      // focus state changes, see [_handleFocusChange].
-    }
-  }
-
-  FocusOwner get focusOwner {
-    assert(!_disposed);
-    if (_focusNode.hasFocus) {
-      return FocusOwner.editor;
-    } else if (_toolbarFocusNode?.hasFocus == true) {
-      return FocusOwner.toolbar;
-    } else {
-      return FocusOwner.none;
-    }
-  }
-
-  void updateSelection(TextSelection value,
-      {ChangeSource source: ChangeSource.remote}) {
-    assert(!_disposed);
-    _controller.updateSelection(value, source: source);
-  }
-
-  void formatSelection(NotusAttribute value) {
-    assert(!_disposed);
-    _controller.formatSelection(value);
-  }
-
-  void focus() {
-    assert(!_disposed);
-    _focusScope.requestFocus(_focusNode);
-  }
-
-  void hideKeyboard() {
-    assert(!_disposed);
-    _focusNode.unfocus();
-  }
-}
-
-class _ZefyrEditorScope extends InheritedWidget {
-  final ZefyrEditorScope scope;
-
-  _ZefyrEditorScope({Key key, Widget child, @required this.scope})
-      : super(key: key, child: child);
-
-  @override
-  bool updateShouldNotify(_ZefyrEditorScope oldWidget) {
-    return oldWidget.scope != scope;
-  }
-}
 
 /// Widget for editing Zefyr documents.
 class ZefyrEditor extends StatefulWidget {
@@ -164,38 +20,62 @@ class ZefyrEditor extends StatefulWidget {
     Key key,
     @required this.controller,
     @required this.focusNode,
-    this.autofocus: true,
-    this.enabled: true,
-    this.padding: const EdgeInsets.symmetric(horizontal: 16.0),
+    this.autofocus = true,
+    this.mode = ZefyrMode.edit,
+    this.padding = const EdgeInsets.symmetric(horizontal: 16.0),
     this.toolbarDelegate,
     this.imageDelegate,
+    this.selectionControls,
     this.physics,
-  }) : super(key: key);
+  })  : assert(mode != null),
+        assert(controller != null),
+        assert(focusNode != null),
+        super(key: key);
 
+  /// Controls the document being edited.
   final ZefyrController controller;
+
+  /// Controls whether this editor has keyboard focus.
   final FocusNode focusNode;
+
+  /// Whether this editor should focus itself if nothing else is already
+  /// focused.
+  ///
+  /// If true, the keyboard will open as soon as this text field obtains focus.
+  /// Otherwise, the keyboard is only shown after the user taps the text field.
+  ///
+  /// Defaults to true. Cannot be null.
   final bool autofocus;
-  final bool enabled;
+
+  /// Editing mode of this editor.
+  final ZefyrMode mode;
+
+  /// Optional delegate for customizing this editor's toolbar.
   final ZefyrToolbarDelegate toolbarDelegate;
+
+  /// Delegate for resolving embedded images.
+  ///
+  /// This delegate is required if embedding images is allowed.
   final ZefyrImageDelegate imageDelegate;
+
+  /// Optional delegate for building the text selection handles and toolbar.
+  ///
+  /// If not provided then platform-specific implementation is used by default.
+  final TextSelectionControls selectionControls;
+
+  /// Controls physics of scrollable editor.
   final ScrollPhysics physics;
 
   /// Padding around editable area.
   final EdgeInsets padding;
 
-  static ZefyrEditorScope of(BuildContext context) {
-    _ZefyrEditorScope widget =
-        context.inheritFromWidgetOfExactType(_ZefyrEditorScope);
-    return widget.scope;
-  }
-
   @override
-  _ZefyrEditorState createState() => new _ZefyrEditorState();
+  _ZefyrEditorState createState() => _ZefyrEditorState();
 }
 
 class _ZefyrEditorState extends State<ZefyrEditor> {
   ZefyrImageDelegate _imageDelegate;
-  ZefyrEditorScope _scope;
+  ZefyrScope _scope;
   ZefyrThemeData _themeData;
   GlobalKey<ZefyrToolbarState> _toolbarKey;
   ZefyrScaffoldState _scaffold;
@@ -214,7 +94,7 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
     _toolbarKey = null;
   }
 
-  Widget buildToolbar(BuildContext) {
+  Widget buildToolbar(BuildContext context) {
     return ZefyrTheme(
       data: _themeData,
       child: ZefyrToolbar(
@@ -231,6 +111,7 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
     } else if (!hasToolbar) {
       showToolbar();
     } else {
+      // TODO: is there a nicer way to do this?
       WidgetsBinding.instance.addPostFrameCallback((_) {
         _toolbarKey?.currentState?.markNeedsRebuild();
       });
@@ -240,17 +121,18 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
   @override
   void initState() {
     super.initState();
-    _imageDelegate = widget.imageDelegate ?? new ZefyrDefaultImageDelegate();
+    _imageDelegate = widget.imageDelegate;
   }
 
   @override
   void didUpdateWidget(ZefyrEditor oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _scope._updateControllerIfNeeded(widget.controller);
-    _scope._updateFocusNodeIfNeeded(widget.focusNode);
+    _scope.mode = widget.mode;
+    _scope.controller = widget.controller;
+    _scope.focusNode = widget.focusNode;
     if (widget.imageDelegate != oldWidget.imageDelegate) {
-      _imageDelegate = widget.imageDelegate ?? new ZefyrDefaultImageDelegate();
-      _scope._updateImageDelegateIfNeeded(_imageDelegate);
+      _imageDelegate = widget.imageDelegate;
+      _scope.imageDelegate = _imageDelegate;
     }
   }
 
@@ -264,7 +146,8 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
         : fallbackTheme;
 
     if (_scope == null) {
-      _scope = ZefyrEditorScope(
+      _scope = ZefyrScope.editable(
+        mode: widget.mode,
         imageDelegate: _imageDelegate,
         controller: widget.controller,
         focusNode: widget.focusNode,
@@ -273,9 +156,7 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
       _scope.addListener(_handleChange);
     } else {
       final focusScope = FocusScope.of(context);
-      if (focusScope != _scope._focusScope) {
-        _scope._focusScope = focusScope;
-      }
+      _scope.focusScope = focusScope;
     }
 
     final scaffold = ZefyrScaffold.of(context);
@@ -297,19 +178,20 @@ class _ZefyrEditorState extends State<ZefyrEditor> {
 
   @override
   Widget build(BuildContext context) {
-    Widget editable = new ZefyrEditableText(
-      controller: widget.controller,
-      focusNode: widget.focusNode,
-      imageDelegate: _imageDelegate,
+    Widget editable = ZefyrEditableText(
+      controller: _scope.controller,
+      focusNode: _scope.focusNode,
+      imageDelegate: _scope.imageDelegate,
+      selectionControls: widget.selectionControls,
       autofocus: widget.autofocus,
-      enabled: widget.enabled,
+      mode: widget.mode,
       padding: widget.padding,
       physics: widget.physics,
     );
 
     return ZefyrTheme(
       data: _themeData,
-      child: _ZefyrEditorScope(
+      child: ZefyrScopeAccess(
         scope: _scope,
         child: editable,
       ),
