@@ -2,42 +2,41 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
-import 'dart:io';
 import 'dart:math' as math;
 import 'dart:ui' as ui;
 
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
 import 'package:notus/notus.dart';
-import 'package:image_picker/image_picker.dart';
 
 import 'editable_box.dart';
 
+/// Provides interface for embedding images into Zefyr editor.
+// TODO: allow configuring image sources and related toolbar buttons.
+@experimental
 abstract class ZefyrImageDelegate<S> {
-  /// Builds image widget for specified [imageSource] and [context].
-  Widget buildImage(BuildContext context, String imageSource);
+  /// Unique key to identify camera source.
+  S get cameraSource;
+
+  /// Unique key to identify gallery source.
+  S get gallerySource;
+
+  /// Builds image widget for specified image [key].
+  ///
+  /// The [key] argument contains value which was previously returned from
+  /// [pickImage] method.
+  Widget buildImage(BuildContext context, String key);
 
   /// Picks an image from specified [source].
   ///
   /// Returns unique string key for the selected image. Returned key is stored
   /// in the document.
+  ///
+  /// Depending on your application returned key may represent a path to
+  /// an image file on user's device, an HTTP link, or an identifier generated
+  /// by a file hosting service like AWS S3 or Google Drive.
   Future<String> pickImage(S source);
-}
-
-class ZefyrDefaultImageDelegate implements ZefyrImageDelegate<ImageSource> {
-  @override
-  Widget buildImage(BuildContext context, String imageSource) {
-    final file = new File.fromUri(Uri.parse(imageSource));
-    final image = new FileImage(file);
-    return new Image(image: image);
-  }
-
-  @override
-  Future<String> pickImage(ImageSource source) async {
-    final file = await ImagePicker.pickImage(source: source);
-    if (file == null) return null;
-    return file.uri.toString();
-  }
 }
 
 class ZefyrImage extends StatefulWidget {
@@ -54,7 +53,7 @@ class ZefyrImage extends StatefulWidget {
 class _ZefyrImageState extends State<ZefyrImage> {
   String get imageSource {
     EmbedAttribute attribute = widget.node.style.get(NotusAttribute.embed);
-    return attribute.value['source'];
+    return attribute.value['source'] as String;
   }
 
   @override
@@ -76,7 +75,7 @@ class _EditableImage extends SingleChildRenderObjectWidget {
 
   @override
   RenderEditableImage createRenderObject(BuildContext context) {
-    return new RenderEditableImage(node: node);
+    return RenderEditableImage(node: node);
   }
 
   @override
@@ -101,7 +100,7 @@ class RenderEditableImage extends RenderBox
   @override
   EmbedNode get node => _node;
   EmbedNode _node;
-  void set node(EmbedNode value) {
+  set node(EmbedNode value) {
     _node = value;
   }
 
@@ -130,16 +129,16 @@ class RenderEditableImage extends RenderBox
     if (local.isCollapsed) {
       final dx = local.extentOffset == 0 ? _childOffset.dx : size.width;
       return [
-        new ui.TextBox.fromLTRBD(
+        ui.TextBox.fromLTRBD(
             dx, 0.0, dx, size.height - kPaddingBottom, TextDirection.ltr),
       ];
     }
 
     final rect = _childRect;
     return [
-      new ui.TextBox.fromLTRBD(
+      ui.TextBox.fromLTRBD(
           rect.left, rect.top, rect.left, rect.bottom, TextDirection.ltr),
-      new ui.TextBox.fromLTRBD(
+      ui.TextBox.fromLTRBD(
           rect.right, rect.top, rect.right, rect.bottom, TextDirection.ltr),
     ];
   }
@@ -151,13 +150,13 @@ class RenderEditableImage extends RenderBox
     if (offset.dx > size.width / 2) {
       position++;
     }
-    return new TextPosition(offset: position);
+    return TextPosition(offset: position);
   }
 
   @override
   TextRange getWordBoundary(TextPosition position) {
     final start = _node.documentOffset;
-    return new TextRange(start: start, end: start + 1);
+    return TextRange(start: start, end: start + 1);
   }
 
   @override
@@ -170,10 +169,10 @@ class RenderEditableImage extends RenderBox
   @override
   Offset getOffsetForCaret(TextPosition position, Rect caretPrototype) {
     final pos = position.offset - node.documentOffset;
-    Offset caretOffset = _childOffset - new Offset(kHorizontalPadding, 0.0);
+    Offset caretOffset = _childOffset - Offset(kHorizontalPadding, 0.0);
     if (pos == 1) {
-      caretOffset = caretOffset +
-          new Offset(_lastChildSize.width + kHorizontalPadding, 0.0);
+      caretOffset =
+          caretOffset + Offset(_lastChildSize.width + kHorizontalPadding, 0.0);
     }
     return caretOffset;
   }
@@ -184,12 +183,12 @@ class RenderEditableImage extends RenderBox
     final localSelection = getLocalSelection(selection);
     assert(localSelection != null);
     if (!localSelection.isCollapsed) {
-      final Paint paint = new Paint()
+      final Paint paint = Paint()
         ..color = selectionColor
         ..style = PaintingStyle.stroke
         ..strokeWidth = 3.0;
-      final rect = new Rect.fromLTWH(
-          0.0, 0.0, _lastChildSize.width, _lastChildSize.height);
+      final rect =
+          Rect.fromLTWH(0.0, 0.0, _lastChildSize.width, _lastChildSize.height);
       context.canvas.drawRect(rect.shift(offset + _childOffset), paint);
     }
   }
@@ -205,12 +204,12 @@ class RenderEditableImage extends RenderBox
   Offset get _childOffset {
     final dx = (size.width - _lastChildSize.width) / 2 + kHorizontalPadding;
     final dy = (size.height - _lastChildSize.height - kPaddingBottom) / 2;
-    return new Offset(dx, dy);
+    return Offset(dx, dy);
   }
 
   Rect get _childRect {
-    return new Rect.fromLTWH(_childOffset.dx, _childOffset.dy,
-        _lastChildSize.width, _lastChildSize.height);
+    return Rect.fromLTWH(_childOffset.dx, _childOffset.dy, _lastChildSize.width,
+        _lastChildSize.height);
   }
 
   @override
@@ -227,8 +226,7 @@ class RenderEditableImage extends RenderBox
       );
       child.layout(childConstraints, parentUsesSize: true);
       _lastChildSize = child.size;
-      size = new Size(
-          constraints.maxWidth, _lastChildSize.height + kPaddingBottom);
+      size = Size(constraints.maxWidth, _lastChildSize.height + kPaddingBottom);
     } else {
       performResize();
     }
