@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:notus/notus.dart';
 
 import 'buttons.dart';
-import 'editor.dart';
+import 'scope.dart';
 import 'theme.dart';
 
 /// List of all button actions supported by [ZefyrToolbar] buttons.
@@ -66,7 +66,7 @@ class ZefyrToolbarScaffold extends StatelessWidget {
     Key key,
     @required this.body,
     this.trailing,
-    this.autoImplyTrailing: true,
+    this.autoImplyTrailing = true,
   }) : super(key: key);
 
   final Widget body;
@@ -88,7 +88,7 @@ class ZefyrToolbarScaffold extends StatelessWidget {
     } else if (autoImplyTrailing) {
       children.add(toolbar.buildButton(context, ZefyrToolbarAction.close));
     }
-    return new Container(
+    return Container(
       constraints: constraints,
       child: Material(color: theme.color, child: Row(children: children)),
     );
@@ -102,12 +102,12 @@ class ZefyrToolbar extends StatefulWidget implements PreferredSizeWidget {
   const ZefyrToolbar({
     Key key,
     @required this.editor,
-    this.autoHide: true,
+    this.autoHide = true,
     this.delegate,
   }) : super(key: key);
 
   final ZefyrToolbarDelegate delegate;
-  final ZefyrEditorScope editor;
+  final ZefyrScope editor;
 
   /// Whether to automatically hide this toolbar when editor loses focus.
   final bool autoHide;
@@ -122,7 +122,7 @@ class ZefyrToolbar extends StatefulWidget implements PreferredSizeWidget {
   ZefyrToolbarState createState() => ZefyrToolbarState();
 
   @override
-  ui.Size get preferredSize => new Size.fromHeight(ZefyrToolbar.kToolbarHeight);
+  ui.Size get preferredSize => Size.fromHeight(ZefyrToolbar.kToolbarHeight);
 }
 
 class _ZefyrToolbarScope extends InheritedWidget {
@@ -165,7 +165,7 @@ class ZefyrToolbarState extends State<ZefyrToolbar>
 
   Future<void> showOverlay(WidgetBuilder builder) async {
     assert(_overlayBuilder == null);
-    final completer = new Completer<void>();
+    final completer = Completer<void>();
     setState(() {
       _overlayBuilder = builder;
       _overlayCompleter = completer;
@@ -187,14 +187,14 @@ class ZefyrToolbarState extends State<ZefyrToolbar>
 
   bool get hasOverlay => _overlayBuilder != null;
 
-  ZefyrEditorScope get editor => widget.editor;
+  ZefyrScope get editor => widget.editor;
 
   @override
   void initState() {
     super.initState();
-    _delegate = widget.delegate ?? new _DefaultZefyrToolbarDelegate();
-    _overlayAnimation = new AnimationController(
-        vsync: this, duration: Duration(milliseconds: 100));
+    _delegate = widget.delegate ?? _DefaultZefyrToolbarDelegate();
+    _overlayAnimation =
+        AnimationController(vsync: this, duration: Duration(milliseconds: 100));
     _selection = editor.selection;
   }
 
@@ -202,7 +202,7 @@ class ZefyrToolbarState extends State<ZefyrToolbar>
   void didUpdateWidget(ZefyrToolbar oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.delegate != oldWidget.delegate) {
-      _delegate = widget.delegate ?? new _DefaultZefyrToolbarDelegate();
+      _delegate = widget.delegate ?? _DefaultZefyrToolbarDelegate();
     }
   }
 
@@ -227,7 +227,7 @@ class ZefyrToolbarState extends State<ZefyrToolbar>
     layers.add(toolbar);
 
     if (hasOverlay) {
-      Widget widget = new Builder(builder: _overlayBuilder);
+      Widget widget = Builder(builder: _overlayBuilder);
       assert(widget != null);
       final overlay = FadeTransition(
         key: _overlayKey,
@@ -259,7 +259,7 @@ class ZefyrToolbarState extends State<ZefyrToolbar>
       buildButton(context, ZefyrToolbarAction.quote),
       buildButton(context, ZefyrToolbarAction.code),
       buildButton(context, ZefyrToolbarAction.horizontalRule),
-      ImageButton(),
+      editor.imageDelegate != null ? ImageButton() : Container(),
     ];
     return buttons;
   }
@@ -275,7 +275,7 @@ class ZefyrButtonList extends StatefulWidget {
 }
 
 class _ZefyrButtonListState extends State<ZefyrButtonList> {
-  final ScrollController _controller = new ScrollController();
+  final ScrollController _controller = ScrollController();
   bool _showLeftArrow = false;
   bool _showRightArrow = false;
 
