@@ -11,30 +11,30 @@ import 'package:zefyr/src/widgets/rich_text.dart';
 import 'package:zefyr/zefyr.dart';
 
 void main() {
-  group('$RenderEditableParagraph', () {
-    final doc = new NotusDocument();
+  group('$RenderZefyrParagraph', () {
+    final doc = NotusDocument();
     doc.insert(0, 'This House Is A Circus');
-    final text = new TextSpan(text: 'This House Is A Circus');
+    final text = TextSpan(text: 'This House Is A Circus');
 
     ZefyrRenderContext renderContext;
-    RenderEditableParagraph p;
+    RenderZefyrParagraph p;
 
     setUp(() {
       WidgetsFlutterBinding.ensureInitialized();
-      renderContext = new ZefyrRenderContext();
-      p = new RenderEditableParagraph(
+      renderContext = ZefyrRenderContext();
+      p = RenderZefyrParagraph(
         text,
-        node: doc.root.children.first,
+        node: doc.root.children.first as LineNode,
         textDirection: TextDirection.ltr,
       );
     });
 
     test('it registers with viewport', () {
-      var owner = new PipelineOwner();
+      var owner = PipelineOwner();
       expect(renderContext.active, isNot(contains(p)));
       p.attach(owner);
       expect(renderContext.dirty, contains(p));
-      p.layout(new BoxConstraints());
+      p.layout(BoxConstraints());
       expect(renderContext.active, contains(p));
     }, skip: 'TODO: move to RenderEditableProxyBox');
   });
