@@ -7,7 +7,7 @@ import 'package:quill_delta/quill_delta.dart';
 import 'package:zefyr/src/widgets/selection.dart';
 import 'package:zefyr/zefyr.dart';
 
-var delta = new Delta()..insert('This House Is A Circus\n');
+var delta = Delta()..insert('This House Is A Circus\n');
 
 class EditorSandBox {
   final WidgetTester tester;
@@ -21,12 +21,20 @@ class EditorSandBox {
     FocusNode focusNode,
     NotusDocument document,
     ZefyrThemeData theme,
+    bool autofocus = false,
+    ZefyrImageDelegate imageDelegate,
   }) {
     focusNode ??= FocusNode();
     document ??= NotusDocument.fromDelta(delta);
     var controller = ZefyrController(document);
 
-    Widget widget = _ZefyrSandbox(controller: controller, focusNode: focusNode);
+    Widget widget = _ZefyrSandbox(
+      controller: controller,
+      focusNode: focusNode,
+      autofocus: autofocus,
+      imageDelegate: imageDelegate,
+    );
+
     if (theme != null) {
       widget = ZefyrTheme(data: theme, child: widget);
     }
@@ -49,7 +57,7 @@ class EditorSandBox {
 
   Future<void> updateSelection({int base, int extent}) {
     controller.updateSelection(
-      new TextSelection(baseOffset: base, extentOffset: extent),
+      TextSelection(baseOffset: base, extentOffset: extent),
     );
     return tester.pumpAndSettle();
   }
@@ -60,11 +68,19 @@ class EditorSandBox {
     return tester.pumpAndSettle();
   }
 
-  Future<void> tapEditor() async {
+  Future<void> pump() async {
     await tester.pumpWidget(widget);
+  }
+
+  Future<void> tap() async {
     await tester.tap(find.byType(ZefyrParagraph).first);
     await tester.pumpAndSettle();
     expect(focusNode.hasFocus, isTrue);
+  }
+
+  Future<void> pumpAndTap() async {
+    await pump();
+    await tap();
   }
 
   Future<void> tapHideKeyboardButton() async {
@@ -96,15 +112,22 @@ class EditorSandBox {
   Finder findSelectionHandle() {
     return find.descendant(
         of: find.byType(SelectionHandleDriver),
-        matching: find.byType(Positioned));
+        matching: find.byType(GestureDetector));
   }
 }
 
 class _ZefyrSandbox extends StatefulWidget {
-  const _ZefyrSandbox({Key key, this.controller, this.focusNode})
-      : super(key: key);
+  const _ZefyrSandbox({
+    Key key,
+    this.controller,
+    this.focusNode,
+    this.autofocus,
+    this.imageDelegate,
+  }) : super(key: key);
   final ZefyrController controller;
   final FocusNode focusNode;
+  final bool autofocus;
+  final ZefyrImageDelegate imageDelegate;
 
   @override
   _ZefyrSandboxState createState() => _ZefyrSandboxState();
@@ -115,10 +138,12 @@ class _ZefyrSandboxState extends State<_ZefyrSandbox> {
 
   @override
   Widget build(BuildContext context) {
-    return new ZefyrEditor(
+    return ZefyrEditor(
       controller: widget.controller,
       focusNode: widget.focusNode,
-      enabled: _enabled,
+      mode: _enabled ? ZefyrMode.edit : ZefyrMode.view,
+      autofocus: widget.autofocus,
+      imageDelegate: widget.imageDelegate,
     );
   }
 
