@@ -8,10 +8,10 @@ import 'package:zefyr/util.dart';
 void main() {
   group('getPositionDelta', () {
     test('actual has more characters inserted than user', () {
-      final user = new Delta()
+      final user = Delta()
         ..retain(7)
         ..insert('a');
-      final actual = new Delta()
+      final actual = Delta()
         ..retain(7)
         ..insert('\na');
       final result = getPositionDelta(user, actual);
@@ -19,10 +19,10 @@ void main() {
     });
 
     test('actual has less characters inserted than user', () {
-      final user = new Delta()
+      final user = Delta()
         ..retain(7)
         ..insert('abc');
-      final actual = new Delta()
+      final actual = Delta()
         ..retain(7)
         ..insert('ab');
       final result = getPositionDelta(user, actual);
@@ -30,10 +30,10 @@ void main() {
     });
 
     test('actual has less characters deleted than user', () {
-      final user = new Delta()
+      final user = Delta()
         ..retain(7)
         ..delete(3);
-      final actual = new Delta()
+      final actual = Delta()
         ..retain(7)
         ..delete(2);
       final result = getPositionDelta(user, actual);
