@@ -1,8 +1,9 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:zefyr/src/widgets/buttons.dart';
 import 'package:zefyr/zefyr.dart';
@@ -12,8 +13,8 @@ import '../testing.dart';
 void main() {
   group('$ZefyrButton', () {
     testWidgets('toggle style', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.updateSelection(base: 5, extent: 10);
       await editor.tapButtonWithIcon(Icons.format_bold);
 
@@ -24,14 +25,14 @@ void main() {
       expect(bold.value, 'House');
 
       await editor.tapButtonWithIcon(Icons.format_bold);
-      line = editor.document.root.children.first;
+      line = editor.document.root.children.first as LineNode;
       expect(line.childCount, 1);
     });
 
     testWidgets('toggle state for different styles of the same attribute',
         (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
 
       await editor.tapButtonWithIcon(Icons.format_list_bulleted);
       expect(editor.document.root.children.first, isInstanceOf<BlockNode>());
@@ -45,8 +46,8 @@ void main() {
 
   group('$HeadingButton', () {
     testWidgets('toggle menu', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_size);
 
       expect(find.text('H1'), findsOneWidget);
@@ -59,8 +60,8 @@ void main() {
     });
 
     testWidgets('toggle styles', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_size);
       await editor.tapButtonWithText('H3');
       LineNode line = editor.document.root.children.first;
@@ -70,8 +71,8 @@ void main() {
     });
 
     testWidgets('close overlay', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_size);
       expect(find.text('H1'), findsOneWidget);
       await editor.tapButtonWithIcon(Icons.close);
@@ -81,24 +82,24 @@ void main() {
 
   group('$LinkButton', () {
     testWidgets('disabled when selection is collapsed', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.link);
       expect(find.byIcon(Icons.link_off), findsNothing);
     });
 
     testWidgets('enabled and toggles menu with non-empty selection',
         (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.updateSelection(base: 5, extent: 10);
       await editor.tapButtonWithIcon(Icons.link);
       expect(find.byIcon(Icons.link_off), findsOneWidget);
     });
 
     testWidgets('auto cancels edit on selection update', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.updateSelection(base: 5, extent: 10);
       await editor.tapButtonWithIcon(Icons.link);
       await tester
@@ -110,8 +111,8 @@ void main() {
     });
 
     testWidgets('editing link', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.updateSelection(base: 5, extent: 10);
 
       await editor.tapButtonWithIcon(Icons.link);
@@ -139,28 +140,18 @@ void main() {
       await editor.updateSelection(base: 7, extent: 7);
       await editor.tapButtonWithIcon(Icons.link);
       await editor.tapButtonWithIcon(Icons.link_off);
-      line = editor.document.root.children.first;
+      line = editor.document.root.children.first as LineNode;
       expect(line.childCount, 1);
     });
   });
 
   group('$ImageButton', () {
-    const MethodChannel channel =
-        const MethodChannel('plugins.flutter.io/image_picker');
-
-    final List<MethodCall> log = <MethodCall>[];
-
-    setUp(() {
-      channel.setMockMethodCallHandler((MethodCall methodCall) async {
-        log.add(methodCall);
-        return '/tmp/test.jpg';
-      });
-      log.clear();
-    });
-
     testWidgets('toggle overlay', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
 
       expect(find.byIcon(Icons.photo_camera), findsOneWidget);
@@ -169,41 +160,43 @@ void main() {
     });
 
     testWidgets('pick from camera', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_camera);
-      expect(log, hasLength(1));
-      expect(
-        log.single,
-        isMethodCall(
-          'pickImage',
-          arguments: <String, dynamic>{
-            'source': 0,
-            'maxWidth': null,
-            'maxHeight': null,
-          },
-        ),
-      );
+      expect(find.byType(ZefyrImage), findsOneWidget);
     });
 
     testWidgets('pick from gallery', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_library);
-      expect(log, hasLength(1));
-      expect(
-        log.single,
-        isMethodCall(
-          'pickImage',
-          arguments: <String, dynamic>{
-            'source': 1,
-            'maxWidth': null,
-            'maxHeight': null,
-          },
-        ),
-      );
+      expect(find.byType(ZefyrImage), findsOneWidget);
     });
   });
+}
+
+class _TestImageDelegate implements ZefyrImageDelegate<String> {
+  @override
+  Widget buildImage(BuildContext context, String key) {
+    return Image.file(File(key));
+  }
+
+  @override
+  String get cameraSource => "camera";
+
+  @override
+  String get gallerySource => "gallery";
+
+  @override
+  Future<String> pickImage(String source) {
+    return Future.value("file:///tmp/test.jpg");
+  }
 }
