@@ -1,53 +1,22 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:zefyr/zefyr.dart';
 
 import '../testing.dart';
 
 void main() {
-  group('$ZefyrDefaultImageDelegate', () {
-    const MethodChannel channel =
-        const MethodChannel('plugins.flutter.io/image_picker');
-
-    final List<MethodCall> log = <MethodCall>[];
-
-    setUp(() {
-      channel.setMockMethodCallHandler((MethodCall methodCall) async {
-        log.add(methodCall);
-        return '/tmp/test.jpg';
-      });
-      log.clear();
-    });
-
-    test('pick image', () async {
-      final delegate = new ZefyrDefaultImageDelegate();
-      final result = await delegate.pickImage(ImageSource.gallery);
-      expect(result, 'file:///tmp/test.jpg');
-    });
-  });
-
   group('$ZefyrImage', () {
-    const MethodChannel channel =
-        const MethodChannel('plugins.flutter.io/image_picker');
-
-    final List<MethodCall> log = <MethodCall>[];
-
-    setUp(() {
-      channel.setMockMethodCallHandler((MethodCall methodCall) async {
-        log.add(methodCall);
-        return '/tmp/test.jpg';
-      });
-      log.clear();
-    });
-
     testWidgets('embed image', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_camera);
       LineNode line = editor.document.root.children.last;
@@ -62,8 +31,11 @@ void main() {
 
     testWidgets('tap on left side of image puts caret before it',
         (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_camera);
       await editor.updateSelection(base: 0, extent: 0);
@@ -75,16 +47,18 @@ void main() {
       expect(editor.selection.extentOffset, embed.documentOffset);
     });
 
-    testWidgets('tap right side of image puts caret after it',
-        (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+    testWidgets('tap right side of image puts caret after it', (tester) async {
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_camera);
       await editor.updateSelection(base: 0, extent: 0);
 
       final img = find.byType(ZefyrImage);
-      final offset = tester.getBottomRight(img) - new Offset(1.0, 1.0);
+      final offset = tester.getBottomRight(img) - Offset(1.0, 1.0);
       await tester.tapAt(offset);
       LineNode line = editor.document.root.children.last;
       EmbedNode embed = line.children.single;
@@ -93,8 +67,11 @@ void main() {
     });
 
     testWidgets('selects on long press', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(
+        tester: tester,
+        imageDelegate: _TestImageDelegate(),
+      );
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.photo);
       await editor.tapButtonWithIcon(Icons.photo_camera);
       await editor.updateSelection(base: 0, extent: 0);
@@ -105,6 +82,25 @@ void main() {
       EmbedNode embed = line.children.single;
       expect(editor.selection.baseOffset, embed.documentOffset);
       expect(editor.selection.extentOffset, embed.documentOffset + 1);
+      expect(find.text('PASTE'), findsOneWidget);
     });
   });
+}
+
+class _TestImageDelegate implements ZefyrImageDelegate<String> {
+  @override
+  Widget buildImage(BuildContext context, String key) {
+    return Image.file(File(key));
+  }
+
+  @override
+  String get cameraSource => "camera";
+
+  @override
+  String get gallerySource => "gallery";
+
+  @override
+  Future<String> pickImage(String source) {
+    return Future.value("file:///tmp/test.jpg");
+  }
 }
