@@ -10,8 +10,8 @@ import '../testing.dart';
 void main() {
   group('$ZefyrList', () {
     testWidgets('format as list', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_list_bulleted);
       BlockNode block = editor.document.root.children.first;
       expect(block.style.get(NotusAttribute.block),
