@@ -10,8 +10,8 @@ import '../testing.dart';
 void main() {
   group('$ZefyrQuote', () {
     testWidgets('format as quote', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_quote);
 
       BlockNode block = editor.document.root.children.first;
