@@ -9,17 +9,17 @@ import 'package:zefyr/src/widgets/rich_text.dart';
 import 'package:zefyr/zefyr.dart';
 
 void main() {
-  group('$EditableRichText', () {
-    final doc = new NotusDocument();
+  group('$ZefyrRichText', () {
+    final doc = NotusDocument();
     doc.insert(0, 'This House Is A Circus');
-    final text = new TextSpan(text: 'This House Is A Circus');
+    final text = TextSpan(text: 'This House Is A Circus');
 
     Widget widget;
     setUp(() {
-      widget = new Directionality(
+      widget = Directionality(
         textDirection: TextDirection.ltr,
-        child: new EditableRichText(
-          node: doc.root.children.first,
+        child: ZefyrRichText(
+          node: doc.root.children.first as LineNode,
           text: text,
         ),
       );
@@ -27,8 +27,7 @@ void main() {
 
     testWidgets('initialize', (tester) async {
       await tester.pumpWidget(widget);
-      EditableRichText result =
-          tester.firstWidget(find.byType(EditableRichText));
+      ZefyrRichText result = tester.firstWidget(find.byType(ZefyrRichText));
       expect(result, isNotNull);
       expect(result.text.text, 'This House Is A Circus');
     });
