@@ -11,23 +11,23 @@ import '../testing.dart';
 void main() {
   group('$ZefyrSelectionOverlay', () {
     testWidgets('double tap caret shows toolbar', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
 
-      RenderEditableParagraph renderObject =
-          tester.firstRenderObject(find.byType(EditableRichText));
+      RenderZefyrParagraph renderObject =
+          tester.firstRenderObject(find.byType(ZefyrRichText));
       var offset = renderObject.localToGlobal(Offset.zero);
       offset += Offset(5.0, 5.0);
       await tester.tapAt(offset);
       await tester.pumpAndSettle();
       await tester.tapAt(offset);
       await tester.pumpAndSettle();
-      expect(find.text('Paste'), findsOneWidget);
+      expect(find.text('PASTE'), findsOneWidget);
     });
 
     testWidgets('hides when editor lost focus', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       await editor.updateSelection(base: 0, extent: 5);
       expect(editor.findSelectionHandle(), findsNWidgets(2));
       await editor.unfocus();
@@ -35,15 +35,14 @@ void main() {
     });
 
     testWidgets('tap on padding area finds closest paragraph', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
-      editor.controller
-          .updateSelection(new TextSelection.collapsed(offset: 10));
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
+      editor.controller.updateSelection(TextSelection.collapsed(offset: 10));
       await tester.pumpAndSettle();
       expect(editor.controller.selection.extentOffset, 10);
 
-      RenderEditableParagraph renderObject =
-          tester.firstRenderObject(find.byType(EditableRichText));
+      RenderZefyrParagraph renderObject =
+          tester.firstRenderObject(find.byType(ZefyrRichText));
       var offset = renderObject.localToGlobal(Offset.zero);
       offset += Offset(-5.0, 5.0);
       await tester.tapAt(offset);
@@ -53,10 +52,10 @@ void main() {
     });
 
     testWidgets('tap on empty space finds closest paragraph', (tester) async {
-      final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      final editor = EditorSandBox(tester: tester);
+      await editor.pumpAndTap();
       editor.controller.replaceText(10, 1, '\n',
-          selection: new TextSelection.collapsed(offset: 0));
+          selection: TextSelection.collapsed(offset: 0));
       await tester.pumpAndSettle();
       expect(editor.controller.document.toPlainText(),
           'This House\nIs A Circus\n');
